% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{arrange.planktonr_dat}
\alias{arrange.planktonr_dat}
\title{Arrange rows of a planktonr_dat object}
\usage{
\method{arrange}{planktonr_dat}(.data, ..., .by_group = FALSE)
}
\arguments{
\item{.data}{A \code{planktonr_dat} object.}

\item{...}{<\code{\link[dplyr:dplyr_data_masking]{data-masking}}> Columns to arrange by.}

\item{.by_group}{If \code{TRUE}, will arrange first by the grouping variables.
See \code{\link[dplyr:arrange]{dplyr::arrange()}} for details.}
}
\value{
A \code{planktonr_dat} object with arranged rows,
preserving original attributes.
}
\description{
This is an S3 method for \code{dplyr::arrange} that ensures
\code{planktonr_dat} attributes are preserved when arranging rows.
}
