% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_indices_nrs.R
\name{create_indices_nrs}
\alias{create_indices_nrs}
\title{Create a range of indices}
\usage{
create_indices_nrs()
}
\value{
A dataframe with with NRS indices
}
\description{
Create a range of indices
}
\examples{
df <- create_indices_nrs()
}
