% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planktonr_dat.R
\name{create_planktonr_dat}
\alias{create_planktonr_dat}
\title{Define the constructor function for the planktonr data class}
\usage{
create_planktonr_dat(data, type, source)
}
\arguments{
\item{data}{The data tibble to be updated to the planktonr class}

\item{type}{The data type - Microbes, Phytoplankton or Zooplankton}

\item{source}{The source of the data - NRS, CPR, LTM}
}
\value{
An object of class planktonr_dat
}
\description{
Define the constructor function for the planktonr data class
}
