% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planktonr_dat.R
\name{drop_na.planktonr_dat}
\alias{drop_na.planktonr_dat}
\title{drop_na method for planktonr_dat}
\usage{
drop_na.planktonr_dat(data, ...)
}
\arguments{
\item{data}{The planktonr data object}

\item{...}{Other possible objects to be passed to the tidyverse}
}
\description{
drop_na method for planktonr_dat
}
