% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_processed_cpr.R
\name{export_processed_cpr}
\alias{export_processed_cpr}
\title{Create all processed data products for CPR}
\usage{
export_processed_cpr(outD)
}
\arguments{
\item{outD}{The directory where the output will be saved.}
}
\description{
Create all processed data products for CPR
}
\examples{

}
