% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_processed_larvalfish.R
\name{export_processed_larvalfish}
\alias{export_processed_larvalfish}
\title{Create all processed data products for Larval Fish}
\usage{
export_processed_larvalfish(outD)
}
\arguments{
\item{outD}{The directory where the output will be saved.}
}
\description{
Create all processed data products for Larval Fish
}
\examples{

}
