% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_processed_nrs.R
\name{export_processed_nrs}
\alias{export_processed_nrs}
\title{Create all processed data products}
\usage{
export_processed_nrs(outD)
}
\arguments{
\item{outD}{The directory where the output will be saved.}
}
\description{
Create all processed data products
}
\examples{

}
