% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{filter.planktonr_dat}
\alias{filter.planktonr_dat}
\title{Filter rows of a planktonr_dat object}
\usage{
\method{filter}{planktonr_dat}(.data, ...)
}
\arguments{
\item{.data}{A \code{planktonr_dat} object.}

\item{...}{<\code{\link[dplyr:dplyr_data_masking]{data-masking}}> Logical predicates to filter by.}
}
\value{
A \code{planktonr_dat} object with filtered rows,
preserving original attributes.
}
\description{
This is an S3 method for \code{dplyr::filter} that ensures
\code{planktonr_dat} attributes are preserved when filtering rows.
}
