% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nrs.R
\name{get_Chemistry}
\alias{get_Chemistry}
\title{Load chemistry data}
\usage{
get_Chemistry()
}
\value{
A dataframe with NRS chemistry data
}
\description{
Load chemistry data
}
\examples{
df <- get_Chemistry()
}
