% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cpr.R
\name{get_CPRHTGBV}
\alias{get_CPRHTGBV}
\title{Get CPR Phyto HTG product - Biovolume}
\usage{
get_CPRHTGBV()
}
\value{

}
\description{
Get CPR Phyto HTG product - Biovolume
}
\examples{
df <- get_CPRHTGBV()
}
