% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cpr.R
\name{get_CPRPhytoGenus}
\alias{get_CPRPhytoGenus}
\title{Get CPR Phyto genus pivoted product - Abundance}
\usage{
get_CPRPhytoGenus()
}
\value{

}
\description{
Get CPR Phyto genus pivoted product - Abundance
}
\examples{
df <- get_CPRPhytoGenus()
}
