% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cpr.R
\name{get_CPRPhytoGenusBV}
\alias{get_CPRPhytoGenusBV}
\title{Get CPR Phyto genus product - Biovolume}
\usage{
get_CPRPhytoGenusBV()
}
\value{

}
\description{
Get CPR Phyto genus product - Biovolume
}
\examples{
df <- get_CPRPhytoGenusBV()
}
