% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cpr.R
\name{get_CPRPhytoSpeciesBV}
\alias{get_CPRPhytoSpeciesBV}
\title{Get CPR Phyto species product - Biovolume}
\usage{
get_CPRPhytoSpeciesBV()
}
\value{

}
\description{
Get CPR Phyto species product - Biovolume
}
\examples{
df <- get_CPRPhytoSpeciesBV()
}
