% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cpr.R
\name{get_CPRZooCopepod}
\alias{get_CPRZooCopepod}
\title{CPR Zoop copepod product - Abundance}
\usage{
get_CPRZooCopepod()
}
\value{

}
\description{
CPR Zoop copepod product - Abundance
}
\examples{
df <- get_CPRZooCopepod()
}
