% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cpr.R
\name{get_CPRZooData}
\alias{get_CPRZooData}
\title{Get CPR Zooplankton abundance data}
\usage{
get_CPRZooData()
}
\value{

}
\description{
Get CPR Zooplankton abundance data
}
\examples{
df <- get_CPRZooData()
}
