% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cpr.R
\name{get_CPRZooGenus}
\alias{get_CPRZooGenus}
\title{CPR Zoop genus product - Abundance}
\usage{
get_CPRZooGenus()
}
\value{

}
\description{
CPR Zoop genus product - Abundance
}
\examples{
df <- get_CPRZooGenus()
}
