% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_larvalfish.R
\name{get_LFCountAll}
\alias{get_LFCountAll}
\title{Make count data of all larval fish}
\usage{
get_LFCountAll()
}
\value{

}
\description{
Make count data of all larval fish
}
\examples{
df <- get_LFCountAll()
}
