% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_larvalfish.R
\name{get_LFCountBGC}
\alias{get_LFCountBGC}
\title{Make BGC data for larval fish}
\usage{
get_LFCountBGC()
}
\value{

}
\description{
Make BGC data for larval fish
}
\examples{
df <- get_LFCountBGC()
}
