% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_larvalfish.R
\name{get_LFData}
\alias{get_LFData}
\title{Get Larval Fish Sample Data}
\usage{
get_LFData()
}
\value{

}
\description{
Get Larval Fish Sample Data
}
\examples{
df <- get_LFData()
#' @importFrom magrittr "\%>\%"
}
