% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_larvalfish.R
\name{get_LFTrips}
\alias{get_LFTrips}
\title{Get Larval Fish Trip Data}
\usage{
get_LFTrips()
}
\value{

}
\description{
Get Larval Fish Trip Data
}
\examples{
df <- get_LFTrips()
#' @importFrom magrittr "\%>\%"
}
