% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nrs.R
\name{get_NRSPhytoData}
\alias{get_NRSPhytoData}
\title{Import NRS Phytoplankton Data}
\usage{
get_NRSPhytoData()
}
\value{
A dataframe with NRS Phytoplankton Data
}
\description{
Load NRS station Phytoplankton Data
}
\examples{
df <- get_NRSPhytoData()
}
