% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nrs.R
\name{get_NRSPhytoGenus}
\alias{get_NRSPhytoGenus}
\title{Get NRS Phytoplankton genus pivoted product - Abundance}
\usage{
get_NRSPhytoGenus()
}
\value{

}
\description{
Get NRS Phytoplankton genus pivoted product - Abundance
}
\examples{
df <- get_NRSPhytoGenus()
}
