% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nrs.R
\name{get_NRSPhytoGenusBV}
\alias{get_NRSPhytoGenusBV}
\title{Get NRS Phytoplankton genus pivoted product - Biovolume}
\usage{
get_NRSPhytoGenusBV()
}
\value{

}
\description{
Get NRS Phytoplankton genus pivoted product - Biovolume
}
\examples{
df <- get_NRSPhytoGenusBV()
}
