% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nrs.R
\name{get_NRSPhytoHTG}
\alias{get_NRSPhytoHTG}
\title{Get Abundance of Phyto Higher Trophic Groups}
\usage{
get_NRSPhytoHTG()

get_NRSPhytoHTG()
}
\value{



}
\description{
Get Abundance of Phyto Higher Trophic Groups

Get NRS Phytoplankton higher taxon group pivoted product - Abundance
}
\examples{
df <- get_NRSPhytoHTG()
#' @importFrom magrittr "\%>\%"
df <- get_NRSPhytoHTG()
}
