% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nrs.R
\name{get_NRSPico}
\alias{get_NRSPico}
\title{Load picophytoplankton data}
\usage{
get_NRSPico()
}
\value{
A dataframe with NRS picophytoplankton data
}
\description{
Load picophytoplankton data
}
\examples{
df <- get_NRSPico()
}
