% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nrs.R
\name{get_NRSPigments}
\alias{get_NRSPigments}
\title{Get pigments data}
\usage{
get_NRSPigments()
}
\value{
A dataframe with NRS pigment data
}
\description{
Get pigments data
}
\examples{
df <- get_NRSPigments()
}
