% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nrs.R
\name{get_NRSStation}
\alias{get_NRSStation}
\title{Import NRS Station information}
\usage{
get_NRSStation()
}
\value{
A dataframe with NRS Station information
}
\description{
Load NRS station information including codes, location, and sampling interval.
}
\examples{
df <- get_NRSStation()
}
