% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nrs.R
\name{get_NRSTrips}
\alias{get_NRSTrips}
\title{Import NRS BGC information}
\usage{
get_NRSTrips()
}
\value{
A dataframe with NRS BGC information
}
\description{
Load NRS station BGC information
}
\examples{
df <- get_NRSTrips()
}
