% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nrs.R
\name{get_NRSZooCount}
\alias{get_NRSZooCount}
\title{Load zooplankton abundance data}
\usage{
get_NRSZooCount()
}
\value{
A dataframe with zooplankton abundance data
}
\description{
Load zooplankton abundance data
}
\examples{
df <- get_NRSZooCount()
}
