% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_raw_plankton}
\alias{get_raw_plankton}
\title{Get location of raw plankton data}
\usage{
get_raw_plankton()
}
\value{
A string with location of raw plankton data
}
\description{
Internal function to load the loaction of the raw plankton data files.
}
\examples{
NRSStation <- get_NRSStation()
}
