% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ZooInfo}
\alias{get_ZooInfo}
\title{Load copepod information table with sizes etc.}
\usage{
get_ZooInfo()
}
\value{
A dataframe with NRS zooplankton information
}
\description{
Load copepod information table with sizes etc.
}
\examples{
df <- get_ZooInfo()
}
