% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getCPRPhytoGenus}
\alias{getCPRPhytoGenus}
\title{Get CPR Phyto genus pivoted product - Abundance}
\usage{
getCPRPhytoGenus()
}
\value{

}
\description{
Get CPR Phyto genus pivoted product - Abundance
}
\examples{
df <- getCPRPhytoGenus()
}
