% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getCPRPhytoGenusBV}
\alias{getCPRPhytoGenusBV}
\title{Get CPR Phyto genus product - Biovolume}
\usage{
getCPRPhytoGenusBV()
}
\value{

}
\description{
Get CPR Phyto genus product - Biovolume
}
\examples{
df <- getCPRPhytoGenusBV()
}
