% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getCPRPhytoSpeciesBV}
\alias{getCPRPhytoSpeciesBV}
\title{Get CPR Phyto species product - Biovolume}
\usage{
getCPRPhytoSpeciesBV()
}
\value{

}
\description{
Get CPR Phyto species product - Biovolume
}
\examples{
df <- getCPRPhytoSpeciesBV()
}
