% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getCPRZooCopepod}
\alias{getCPRZooCopepod}
\title{CPR Zoop copepod product - Abundance}
\usage{
getCPRZooCopepod()
}
\value{

}
\description{
CPR Zoop copepod product - Abundance
}
\examples{
df <- getCPRZooCopepod()
}
