% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getCPRZooRawSS}
\alias{getCPRZooRawSS}
\title{CPR Zoop raw product binned by sex and stage raw pivoted product}
\usage{
getCPRZooRawSS()
}
\value{

}
\description{
CPR Zoop raw product binned by sex and stage raw pivoted product
}
\examples{
df <- getCPRZooRawSS()
}
