% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getNRSPhytoData}
\alias{getNRSPhytoData}
\title{Import NRS Phytoplankton Data}
\usage{
getNRSPhytoData()
}
\value{
A dataframe with NRS Phytoplankton Data
}
\description{
Load NRS station Phytoplankton Data
}
\examples{
df <- getNRSPhytoData()
}
