% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getNRSStation}
\alias{getNRSStation}
\title{Import NRS Station information}
\usage{
getNRSStation()
}
\value{
A dataframe with NRS Station information
}
\description{
Load NRS station information including codes, location, and sampling interval.
}
\examples{
df <- getNRSStation()
}
