% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getNRSTrips}
\alias{getNRSTrips}
\title{Import NRS BGC information}
\usage{
getNRSTrips()
}
\value{
A dataframe with NRS BGC information
}
\description{
Load NRS station BGC information
}
\examples{
df <- getNRSTrips()
}
