% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getNRSZooCount}
\alias{getNRSZooCount}
\title{Load zooplankton abundance data}
\usage{
getNRSZooCount()
}
\value{
A dataframe with zooplankton abundance data
}
\description{
Load zooplankton abundance data
}
\examples{
df <- getNRSZooCount()
}
