% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getNRSZooData}
\alias{getNRSZooData}
\title{Load zooplankton abundance data}
\usage{
getNRSZooData()
}
\value{
A dataframe with zooplankton abundance data
}
\description{
Load zooplankton abundance data
}
\examples{
df <- getNRSZooData()
}
