% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getPico}
\alias{getPico}
\title{Load picophytoplankton data}
\usage{
getPico()
}
\value{
A dataframe with NRS picophytoplankton data
}
\description{
Load picophytoplankton data
}
\examples{
df <- getPico()
}
