% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getZooInfo}
\alias{getZooInfo}
\title{Load copepod information table with sizes etc.}
\usage{
getZooInfo()
}
\value{
A dataframe with NRS zooplankton information
}
\description{
Load copepod information table with sizes etc.
}
\examples{
df <- getZooInfo()
}
