% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MapOz}
\alias{MapOz}
\title{MapOz.}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 64 columns.
}
\usage{
MapOz
}
\description{
A dataset for plotting the base Australia map.
}
\keyword{datasets}
