% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{\%m+\%}
\alias{\%m+\%}
\title{Pipe operator for adding time periods to dates}
\usage{
e1 \%m+\% e2
}
\arguments{
\item{e1}{A period or a date-time object of class.}

\item{e2}{A period or a date-time object of class.}
}
\value{
A date-time object of class POSIXlt, POSIXct or Date.
}
\description{
Pipe operator for adding time periods to dates
}
\keyword{internal}
