% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{\%m+\%}
\alias{\%m+\%}
\title{Pipe operator}
\usage{
lhs \%m+\% rhs
}
\arguments{
\item{lhs}{A period or a date-time object of class.}

\item{rhs}{A period or a date-time object of class.}
}
\value{
The result of calling \code{rhs(lhs)}.
}
\description{
Pipe operator
}
\keyword{internal}
