% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planktonr_dat.R
\name{planktonr_dat}
\alias{planktonr_dat}
\title{Define the constructor function for the planktonr data class}
\usage{
planktonr_dat(
  .data,
  type = NULL,
  survey = NULL,
  variable = NULL,
  model = NULL,
  ...
)
}
\arguments{
\item{.data}{The data tibble to be converted to the \code{planktonr_dat} class.}

\item{type}{The data type. Must be one of "Microbes", "Phytoplankton",
"Zooplankton", "Water", or "EOV".}

\item{survey}{The survey of the data. E.g., "NRS", "CPR", "LTM", "GO-SHIP", "Coastal".}

\item{variable}{What variable is being described or subsetted by this data.}

\item{model}{Optional model associated with the data.}

\item{...}{Additional attributes to be stored with the \code{planktonr_dat} object.
These will be stored as attributes on the object.}
}
\value{
An object of class \code{planktonr_dat}.
}
\description{
This function creates a \code{planktonr_dat} object, which is a tibble
with additional custom attributes for metadata.
}
\examples{
data_tibble <- tibble::tibble(
  time = as.Date(c("2023-01-01", "2023-01-02")),
  value = c(10, 20)
)
pr_obj <- planktonr_dat(
  .data = data_tibble,
  type = "Phytoplankton",
  survey = "NRS",
  variable = "Chlorophyll_a",
  notes = "Example data set"
)
print(pr_obj)
}
