% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pr_add_Carbon}
\alias{pr_add_Carbon}
\title{Add Carbon concentration to phytoplankton dataframe}
\usage{
pr_add_Carbon(df, meth)
}
\arguments{
\item{df}{Input dataframe with BioVolume}

\item{meth}{Method for data collection}
}
\value{
Dataframe with Carbon included
}
\description{
This is where you write the description
}
\examples{
df <- data.frame(TaxonGroup = c("Dinoflagellate", "Cyanobacteria"),
                          Biovolume_um3L = c(100, 150), Cells_L = c(10, 8))
df <- pr_add_Carbon(df, "NRS")
df <- data.frame(TaxonGroup = c("Dinoflagellate", "Cyanobacteria"),
                          BioVolume_um3m3 = c(100, 150), PhytoAbund_m3 = c(10, 8))
df <- pr_add_Carbon(df, "CPR")
}
