% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pr_add_LocalTime}
\alias{pr_add_LocalTime}
\title{Add local Time}
\usage{
pr_add_LocalTime(df)
}
\arguments{
\item{df}{Dataframe with UTC Time}
}
\value{
A datafrane with local time added
}
\description{
This function adds the local time based on timezone.
}
\examples{
df <- data.frame(tz = c("Australia/Perth", "Australia/Brisbane"),
               SampleTime_UTC = c(lubridate::now(), lubridate::now()))
df <- pr_add_LocalTime(df)
}
