% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pr_add_StationName}
\alias{pr_add_StationName}
\title{Add NRS StationName to data}
\usage{
pr_add_StationName(df)
}
\arguments{
\item{df}{A dataframe that contains \code{StationCode} but no \code{StationName}}
}
\value{
A dataframe with StationName added
}
\description{
Add NRS StationName to data
}
\examples{
df <- pr_get_Stations() \%>\%
    pr_add_StationName()
}
