% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planktonr_dat.R
\name{pr_check_type}
\alias{pr_check_type}
\title{Check and update Type as required}
\usage{
pr_check_type(Type)
}
\arguments{
\item{Type}{Data type}
}
\value{
Type as a standardised string
}
\description{
Check and update Type as required
}
