% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_export_cpr.R
\name{pr_export_cpr}
\alias{pr_export_cpr}
\title{Create all processed data products for CPR}
\usage{
pr_export_cpr(outD)
}
\arguments{
\item{outD}{The directory where the output will be saved.}
}
\description{
Create all processed data products for CPR
}
\examples{
\dontrun{
pr_export_cpr("Output")
}
}
