% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_export_larvalfish.R
\name{pr_export_larvalfish}
\alias{pr_export_larvalfish}
\title{Create all processed data products for Larval Fish}
\usage{
pr_export_larvalfish(outD)
}
\arguments{
\item{outD}{The directory where the output will be saved.}
}
\description{
Create all processed data products for Larval Fish
}
\examples{
\dontrun{
pr_export_larvalfish("Output")
}
}
