% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_export_nrs.R
\name{pr_export_nrs}
\alias{pr_export_nrs}
\title{Create all processed data products}
\usage{
pr_export_nrs(outD)
}
\arguments{
\item{outD}{The directory where the output will be saved.}
}
\description{
Create all processed data products
}
\examples{

}
