% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_boo_data.R
\name{pr_export_SppCount}
\alias{pr_export_SppCount}
\title{Summarise the plankton observations}
\usage{
pr_export_SppCount()
}
\value{
a dataframe with a species summary
}
\description{
Summarise the plankton observations from the NRS and CPR.
}
\examples{
df <- pr_export_SppCount()
}
