% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nrs.R
\name{pr_filter_NRSStations}
\alias{pr_filter_NRSStations}
\title{Filter dataframe on \code{StationCode} to return only NRS stations}
\usage{
pr_filter_NRSStations(dat)
}
\arguments{
\item{dat}{Dataframe to filter}
}
\value{
Dataframe with only NRS Stations
}
\description{
Filter dataframe on \code{StationCode} to return only NRS stations
}
\examples{
\dontrun{dat <- pr_filter_NRSStations(dat)}
}
