% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pr_filter_species}
\alias{pr_filter_species}
\title{Remove incorrect species names from dataframe}
\usage{
pr_filter_species(df)
}
\arguments{
\item{df}{A dataframe with species names}
}
\value{
A dataframe with all correct species names
}
\description{
Remove incorrect species names from dataframe
}
\examples{
df <- data.frame(Species = c("IncorrectSpecies cf.", "CorrectSpecies1", NA,
              "CorrectSpecies2", "Incorrect spp., Incorrect/Species"))
df <- pr_filter_species(df)
}
