% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_get_chemistry.R
\name{pr_get_chemistry}
\alias{pr_get_chemistry}
\title{Load chemistry data}
\usage{
pr_get_chemistry()
}
\value{
A dataframe with NRS chemistry data
}
\description{
Load chemistry data
}
\examples{
df <- pr_get_chemistry()
}
