% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cpr.R
\name{pr_get_CPRData}
\alias{pr_get_CPRData}
\title{Import CPR Data}
\usage{
pr_get_CPRData(file)
}
\arguments{
\item{file}{Filename to retrieve from AODN.}
}
\value{
A dataframe with requested plankton data in long form
}
\description{
Load CPR Data
}
\examples{
df <- pr_get_CPRData("cpr_zooplankton_abundance_copepods_data")
}
