% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cpr.R
\name{pr_get_CPRPhytoGenus}
\alias{pr_get_CPRPhytoGenus}
\title{Get CPR Phyto genus pivoted product - Abundance}
\usage{
pr_get_CPRPhytoGenus()
}
\value{

}
\description{
Get CPR Phyto genus pivoted product - Abundance
}
\examples{
df <- pr_get_CPRPhytoGenus()
}
