% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cpr.R
\name{pr_get_CPRPhytoGenusBV}
\alias{pr_get_CPRPhytoGenusBV}
\title{Get CPR Phyto genus product - Biovolume}
\usage{
pr_get_CPRPhytoGenusBV()
}
\value{

}
\description{
Get CPR Phyto genus product - Biovolume
}
\examples{
df <- pr_get_CPRPhytoGenusBV()
}
