% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cpr.R
\name{pr_get_CPRPhytoHTG}
\alias{pr_get_CPRPhytoHTG}
\title{CPR Phyto HTG pivoted product - Abundance}
\usage{
pr_get_CPRPhytoHTG()
}
\value{
A dataframe with CPR Phytoplankton Abundance - Summed by Higher Trophic Groups
}
\description{
CPR Phyto HTG pivoted product - Abundance
}
\examples{
df <- pr_get_CPRPhytoHTG
}
