% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cpr.R
\name{pr_get_CPRPhytoHTGBV}
\alias{pr_get_CPRPhytoHTGBV}
\title{Get CPR Phyto HTG product - Biovolume}
\usage{
pr_get_CPRPhytoHTGBV()
}
\value{
A dataframe with CPR Phytoplankton BioVolume - Summed by Higher Trophic Levels
}
\description{
Get CPR Phyto HTG product - Biovolume
}
\examples{
df <- pr_get_CPRPhytoHTGBV()
}
