% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cpr.R
\name{pr_get_CPRPhytoSpecies}
\alias{pr_get_CPRPhytoSpecies}
\title{Get CPR Phyto species pivoted product - Abundance}
\usage{
pr_get_CPRPhytoSpecies()
}
\value{
A dataframe with CPR Phytoplankton Abundance - Summed by Species
}
\description{
Get CPR Phyto species pivoted product - Abundance
}
\examples{
df <- pr_get_CPRPhytoSpecies()
}
