% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cpr.R
\name{pr_get_CPRPhytoSpeciesBV}
\alias{pr_get_CPRPhytoSpeciesBV}
\title{Get CPR Phyto species product - Biovolume}
\usage{
pr_get_CPRPhytoSpeciesBV()
}
\value{
A dataframe with CPR Phytoplankton BioVolume - Summed by Species
}
\description{
Get CPR Phyto species product - Biovolume
}
\examples{
df <- pr_get_CPRPhytoSpeciesBV()
}
