% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cpr.R
\name{pr_get_CPRSamps}
\alias{pr_get_CPRSamps}
\title{Get CPR samples}
\usage{
pr_get_CPRSamps(...)
}
\arguments{
\item{...}{to allow use of join when used within another function}
}
\value{
A dataframe with CPR Samples
}
\description{
Get CPR samples
}
\examples{
df <- pr_get_CPRSamps(join = "st_nearest_feature")
df <- pr_get_CPRSamps()

#' @importFrom rlang .data
}
