% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cpr.R
\name{pr_get_CPRTrips}
\alias{pr_get_CPRTrips}
\title{Get CPR trips}
\usage{
pr_get_CPRTrips(...)
}
\arguments{
\item{...}{ability to choose join for bioregions}
}
\value{
A dataframe with CPR Trips
}
\description{
Get CPR trips
}
\examples{
df <- pr_get_CPRTrips()
df <- pr_get_CPRTrips(near_dist_km = 250)
}
