% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cpr.R
\name{pr_get_CPRZooChangeLog}
\alias{pr_get_CPRZooChangeLog}
\title{Get CPR zooplankton Change Log}
\usage{
pr_get_CPRZooChangeLog()
}
\value{
A dataframe with the CPR Zooplankton Changelog
}
\description{
Get CPR zooplankton Change Log
}
\examples{
df <- pr_get_CPRZooChangeLog()
}
