% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cpr.R
\name{pr_get_CPRZooCopepod}
\alias{pr_get_CPRZooCopepod}
\title{CPR Zoop copepod product - Abundance}
\usage{
pr_get_CPRZooCopepod()
}
\value{
A dataframe with CPR Copepod Abundance - Summed by Species
}
\description{
CPR Zoop copepod product - Abundance
}
\examples{
df <- pr_get_CPRZooCopepod()
}
