% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cpr.R
\name{pr_get_CPRZooGenus}
\alias{pr_get_CPRZooGenus}
\title{CPR Zoop genus product - Abundance}
\usage{
pr_get_CPRZooGenus()
}
\value{
A dataframe with CPR Zooplankton Abundance - Summed by Genus
}
\description{
CPR Zoop genus product - Abundance
}
\examples{
df <- pr_get_CPRZooGenus()
}
