% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cpr.R
\name{pr_get_CPRZooHTG}
\alias{pr_get_CPRZooHTG}
\title{Get CPR Zoop HTG pivoted product - Abundance}
\usage{
pr_get_CPRZooHTG()
}
\value{
A dataframe with CPR Zooplankton Abundance - Summed by Higher Trophic Levels
}
\description{
Get CPR Zoop HTG pivoted product - Abundance
}
\examples{
df <- pr_get_CPRZooHTG()
}
