% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cpr.R
\name{pr_get_CPRZooRawSS}
\alias{pr_get_CPRZooRawSS}
\title{CPR Zoop raw product binned by sex and stage raw pivoted product}
\usage{
pr_get_CPRZooRawSS()
}
\value{

}
\description{
CPR Zoop raw product binned by sex and stage raw pivoted product
}
\examples{
df <- pr_get_CPRZooRawSS()
}
