% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_bgc.R
\name{pr_get_CSChem}
\alias{pr_get_CSChem}
\title{Load chemistry data}
\usage{
pr_get_CSChem()
}
\value{
A dataframe with NRS chemistry data
}
\description{
Load chemistry data
}
\examples{
df <- pr_get_CSChem()
}
