% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pr_get_csv}
\alias{pr_get_csv}
\title{Download and write csv file}
\usage{
pr_get_csv(file, di)
}
\arguments{
\item{file}{The name of the file to download.}

\item{di}{The directory to save it to}
}
\description{
Function to download plankton csv files from the AODN. Options are:
}
\details{
cpr_derived_indices_data
nrs_derived_indices_data

bgc_chemistry_data,
bgc_pigments_data,
bgc_picoplankton_data,
bgc_tss_data,

cpr_phytoplankton_abundance_raw_data
cpr_phytoplankton_abundance_htg_data
cpr_phytoplankton_abundance_genus_data
cpr_phytoplankton_abundance_species_data

cpr_phytoplankton_biovolume_raw_data
cpr_phytoplankton_biovolume_htg_data
cpr_phytoplankton_biovolume_genus_data
cpr_phytoplankton_biovolume_species_data

cpr_zooplankton_abundance_copepods_data
cpr_zooplankton_abundance_non_copepods_data
cpr_zooplankton_abundance_genus_data
cpr_zooplankton_abundance_htg_data
cpr_zooplankton_abundance_raw_data
}
\examples{
pr_get_csv("bgc_chemistry_data", "~/Downloads")
}
