% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plankton.R
\name{pr_get_CTI}
\alias{pr_get_CTI}
\title{Calculate Community Temperature Index (CTI) for plankton samples}
\usage{
pr_get_CTI(Type = "Zooplankton")
}
\arguments{
\item{Type}{Zooplankton (Copepods) or phytoplankton species}
}
\value{
df of CTI for each sample
}
\description{
Calculate Community Temperature Index (CTI) for plankton samples
}
\examples{
df <- pr_get_CTI("Phytoplankton")

}
