% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_boo_data.R
\name{pr_get_facts}
\alias{pr_get_facts}
\title{Random facts about plankton}
\usage{
pr_get_facts()
}
\value{
A random plankton fact as a string
}
\description{
This function randomly returns a fun fact about plankton.
}
\examples{
pr_get_facts()
}
