% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_boo_data.R
\name{pr_get_fg}
\alias{pr_get_fg}
\title{Get functional group data}
\usage{
pr_get_fg(Survey = "NRS", Type = "Z")
}
\arguments{
\item{Survey}{CPR or NRS data}

\item{Type}{Zooplankton or phytoplankton data}
}
\value{
dataframe for plotting functional group time series info
}
\description{
Get functional group data
}
\examples{
NRSfgz <- planktonr::pr_get_fg("NRS", "Z")
NRSfgp <- planktonr::pr_get_fg("NRS", "P")
CPRfgz <- planktonr::pr_get_fg("CPR", "Z")
CPRfgp <- planktonr::pr_get_fg("CPR", "P")
}
