% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_boo_data.R
\name{pr_get_fMap_data}
\alias{pr_get_fMap_data}
\title{Get data for frequency map plots}
\usage{
pr_get_fMap_data(Type = "Z")
}
\arguments{
\item{Type}{Phytoplankton (P) or Zooplankton (Z), defaults to phyto}
}
\value{
dataframe for plotting with pr_plot_fmap
}
\description{
Get data for frequency map plots
}
\examples{
df <- pr_get_fMap_data("P")
df <- pr_get_fMap_data("Z")
}
