% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plotting.R
\name{pr_get_FreqMap}
\alias{pr_get_FreqMap}
\title{Get data for frequency map plots}
\usage{
pr_get_FreqMap(Type = "Zooplankton")
}
\arguments{
\item{Type}{Phytoplankton (P) or Zooplankton (Z), defaults to phyto}
}
\value{
dataframe for plotting with pr_plot_FreqMap
}
\description{
Get data for frequency map plots
}
\examples{
df <- pr_get_FreqMap("Zooplankton")
}
