% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_get_indices_nrs.R
\name{pr_get_indices_nrs}
\alias{pr_get_indices_nrs}
\title{Create a range of indices}
\usage{
pr_get_indices_nrs()
}
\value{
A dataframe with with NRS indices
}
\description{
Create a range of indices
}
\examples{
df <- pr_get_indices_nrs()
}
