% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_larvalfish.R
\name{pr_get_LFCountBGC}
\alias{pr_get_LFCountBGC}
\title{Make BGC data for larval fish}
\usage{
pr_get_LFCountBGC()
}
\value{

}
\description{
Make BGC data for larval fish
}
\examples{
df <- pr_get_LFCountBGC()
}
