% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_larvalfish.R
\name{pr_get_LFTrips}
\alias{pr_get_LFTrips}
\title{Get Larval Fish Trip Data}
\usage{
pr_get_LFTrips()
}
\value{

}
\description{
Get Larval Fish Trip Data
}
\examples{
df <- pr_get_LFTrips()
#' @importFrom magrittr "\%>\%"
}
