% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_BooPlot.R
\name{pr_get_MapOz}
\alias{pr_get_MapOz}
\title{To product the base map of Australia for plotting}
\usage{
pr_get_MapOz()
}
\value{
an sf object for plotting the base map of Australia
}
\description{
To product the base map of Australia for plotting
}
\examples{
MapOz <- pr_get_MapOz()
}
