% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planktonr_dat.R
\name{pr_get_model}
\alias{pr_get_model}
\alias{pr_get_model.planktonr_dat}
\title{Extract fitted model objects from planktonr data with trend analysis}
\usage{
pr_get_model(x)

\method{pr_get_model}{planktonr_dat}(x)
}
\arguments{
\item{x}{A \code{planktonr_dat} object.}
}
\value{
The \code{Model} attribute from the object.
}
\description{
Extract fitted model objects from planktonr data with trend analysis
}
