% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pr_get_nonTaxaColumns}
\alias{pr_get_nonTaxaColumns}
\title{Return columns that are not taxonomic}
\usage{
pr_get_nonTaxaColumns(Survey = "NRS", Type = "Z")
}
\arguments{
\item{Survey}{\code{NRS} or \code{CPR}}

\item{Type}{\code{Z} or \code{P}}
}
\value{
a string vector of column name
}
\description{
Return columns that are not taxonomic
}
\examples{
str <- pr_get_nonTaxaColumns(Survey = "NRS", Type = "Z")
}
