% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_bgc.R
\name{pr_get_NRSEnvContour}
\alias{pr_get_NRSEnvContour}
\title{Data for environmental contour plots}
\usage{
pr_get_NRSEnvContour(Data = "Chemistry")
}
\arguments{
\item{Data}{Use Chemistry, Pico or Micro, the depth stratified samples}
}
\value{
A dataframe to be used with pr_plot_NRSEnvContour
}
\description{
Data for environmental contour plots
}
\examples{
df <- pr_get_NRSEnvContour('Micro')
}
