% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_bgc.R
\name{pr_get_NRSMicro}
\alias{pr_get_NRSMicro}
\title{Load microbial abundance data from flow cytometry}
\usage{
pr_get_NRSMicro(Survey = "NRS")
}
\arguments{
\item{Survey}{NRS, Coastal or GO-SHIP stations}
}
\value{
A dataframe with NRS microbial data
}
\description{
Load microbial abundance data from flow cytometry
}
\examples{
dat <- pr_get_NRSMicro(Survey = "GO-SHIP")
}
