% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nrs.R
\name{pr_get_NRSPhytoData}
\alias{pr_get_NRSPhytoData}
\title{Import NRS Phytoplankton Data}
\usage{
pr_get_NRSPhytoData()
}
\value{
A dataframe with NRS Phytoplankton Data in long form
}
\description{
Load NRS station Phytoplankton Data
}
\examples{
df <- pr_get_NRSPhytoData()
}
