% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nrs.R
\name{pr_get_NRSPhytoGenus}
\alias{pr_get_NRSPhytoGenus}
\title{Get NRS Phytoplankton genus product - Abundance}
\usage{
pr_get_NRSPhytoGenus()
}
\value{
A dataframe with NRS Phytoplankton Abundance - Summed by Genus
}
\description{
Get NRS Phytoplankton genus product - Abundance
}
\examples{
df <- pr_get_NRSPhytoGenus()
}
