% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nrs.R
\name{pr_get_NRSPhytoHTG}
\alias{pr_get_NRSPhytoHTG}
\title{Get Abundance of Phyto Higher Trophic Groups}
\usage{
pr_get_NRSPhytoHTG()

pr_get_NRSPhytoHTG()
}
\value{



}
\description{
Get Abundance of Phyto Higher Trophic Groups

Get NRS Phytoplankton higher taxon group pivoted product - Abundance
}
\examples{
df <- pr_get_NRSPhytoHTG()
#' @importFrom magrittr "\%>\%"
df <- pr_get_NRSPhytoHTG()
}
