% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nrs.R
\name{pr_get_NRSPhytoRawBV}
\alias{pr_get_NRSPhytoRawBV}
\title{Get NRS Phytoplankton raw pivoted product - Biovolume}
\usage{
pr_get_NRSPhytoRawBV()
}
\value{
A dataframe with raw NRS Phytoplankton BioVolume
}
\description{
Get NRS Phytoplankton raw pivoted product - Biovolume
}
\examples{
df <- pr_get_NRSPhytoRawBV()
}
