% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nrs.R
\name{pr_get_NRSStation}
\alias{pr_get_NRSStation}
\title{Import NRS Station information}
\usage{
pr_get_NRSStation()
}
\value{
A dataframe with NRS Station information
}
\description{
Load NRS station information including codes, location, and sampling interval.
}
\examples{
df <- pr_get_NRSStation()
}
