% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{pr_get_NRSTSS}
\alias{pr_get_NRSTSS}
\title{Load NRS Total Suspended Solids data}
\usage{
pr_get_NRSTSS()
}
\value{
A dataframe with NRS TSS data.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{pr_get_NRSTSS()} was deprecated in planktonr 0.7.0 in favour of
\code{\link[=pr_get_data]{pr_get_data()}} which provides a unified interface for all data types.
}
\examples{
\dontrun{
# Instead of:
dat <- pr_get_NRSTSS()

# Use:
dat <- pr_get_data(Survey = "NRS", Type = "TSS")
}
}
\seealso{
\code{\link[=pr_get_data]{pr_get_data()}} for the preferred interface
}
\keyword{internal}
