% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nrs.R
\name{pr_get_NRSZooGenus}
\alias{pr_get_NRSZooGenus}
\title{NRS Zoop genus product - Abundance}
\usage{
pr_get_NRSZooGenus()
}
\value{
A dataframe with NRS Zooplankton Abundance - Summed by Genus
}
\description{
NRS Zoop genus product - Abundance
}
\examples{
df <- pr_get_NRSZooGenus()
}
