% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nrs.R
\name{pr_get_NRSZooRawBin}
\alias{pr_get_NRSZooRawBin}
\title{NRS Zoop raw product binned by sex and stage raw product}
\usage{
pr_get_NRSZooRawBin()
}
\value{
A dataframe with NRS Zooplankton Abundance - Binned by sex and stage
}
\description{
NRS Zoop raw product binned by sex and stage raw product
}
\examples{
df <- pr_get_NRSZooRawBin()
}
