% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nrs.R
\name{pr_get_NRSZooSpeciesCopepod}
\alias{pr_get_NRSZooSpeciesCopepod}
\title{NRS Zoop copepod product - Abundance}
\usage{
pr_get_NRSZooSpeciesCopepod()
}
\value{
A dataframe with NRS Copepod Abundance - Summed by Species
}
\description{
NRS Zoop copepod product - Abundance
}
\examples{
df <- pr_get_NRSZooSpeciesCopepod()
}
