% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_boo_data.R
\name{pr_get_nuts}
\alias{pr_get_nuts}
\title{Get NRS nutrient timeseries data}
\usage{
pr_get_nuts()
}
\value{
dataframe for plotting nutrient time series info
}
\description{
Get NRS nutrient timeseries data
}
\examples{
df <- pr_get_nuts()
}
