% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_boo_data.R
\name{pr_get_pico}
\alias{pr_get_pico}
\title{Get picoplankton time series data}
\usage{
pr_get_pico()
}
\value{
Dataframe for plotting picoplankton time series
}
\description{
Get picoplankton time series data
}
\examples{
df <- pr_get_pico()
}
