% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_boo_data.R
\name{pr_get_pigs}
\alias{pr_get_pigs}
\title{Get NRS pigment timeseries data}
\usage{
pr_get_pigs()
}
\value{
dataframe for plotting pigment time series info
}
\description{
Get NRS pigment timeseries data
}
\examples{
df <- pr_get_pigs()
}
