% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pr_get_PlanktonInfo}
\alias{pr_get_PlanktonInfo}
\title{Load taxonomic information and trait data for plankton species}
\usage{
pr_get_PlanktonInfo(Type = "Zooplankton")
}
\arguments{
\item{Type}{The type of plankton: "Phytoplankton" or "Zooplankton"
(also accepts "P", "Z", "Phyto", "Zoop")}
}
\value{
A dataframe with taxonomic and trait information.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{pr_get_PlanktonInfo()} was deprecated in planktonr 0.7.0 in favour of
\code{\link[=pr_get_info]{pr_get_info()}} which provides a unified interface for both plankton
taxonomic information and survey policy information.
}
\examples{
\dontrun{
# Instead of:
zoo_info <- pr_get_PlanktonInfo(Type = "Zooplankton")

# Use:
zoo_info <- pr_get_info(Source = "Zooplankton")
}
}
\seealso{
\code{\link[=pr_get_info]{pr_get_info()}} for the preferred interface
}
\keyword{internal}
