% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_policy.R
\name{pr_get_pol}
\alias{pr_get_pol}
\title{Get Policy data set}
\usage{
pr_get_pol(Survey = "NRS")
}
\arguments{
\item{Survey}{"NRS" or "CPR"}
}
\value{
A dataframe with policy data
}
\description{
Get Policy data set
}
\examples{
dat <- pr_get_pol("NRS")
}
