% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_policy.R
\name{pr_get_PolicyData}
\alias{pr_get_PolicyData}
\title{Get Policy data set}
\usage{
pr_get_PolicyData(Survey = "NRS", ...)
}
\arguments{
\item{Survey}{"NRS" or "CPR" or "LTM"}
}
\value{
A dataframe with policy data
}
\description{
Get Policy data set
}
\examples{
df <- pr_get_PolicyData("CPR", join = 'st_nearest_feature')
}
