% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_policy.R
\name{pr_get_polInfo}
\alias{pr_get_polInfo}
\title{Get policy information}
\usage{
pr_get_polInfo(Survey = "NRS")
}
\arguments{
\item{Survey}{"NRS" or "CPR"}
}
\value{

}
\description{
Get policy information
}
\examples{
df <- pr_get_polInfo("NRS")
}
