% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_boo_data.R
\name{pr_get_ProgressMap}
\alias{pr_get_ProgressMap}
\title{Data for IMOS progress map}
\usage{
pr_get_ProgressMap(Survey = c("NRS", "CPR", "Both"))
}
\arguments{
\item{Survey}{one of NRS, CPR, Both}
}
\value{
A dataframe for input into pr_plot_Progress()
}
\description{
Data for IMOS progress map
}
\examples{
df <- pr_get_ProgressMap("Both")
}
