% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plotting.R
\name{pr_get_ProgressMapData}
\alias{pr_get_ProgressMapData}
\title{Prepare sampling coverage data for IMOS progress map visualisation}
\usage{
pr_get_ProgressMapData(Survey = c("NRS", "CPR"), interactive = FALSE, ...)
}
\arguments{
\item{Survey}{one of NRS, CPR or Both}

\item{interactive}{A logical TRUE/FALSE if the data is to be used for an interactive plot.}

\item{...}{variables to be passed to pr_add_Bioregions. At the moment it only supports \code{near_dist_km} which is the distance (km) around each bioregion to pad the allocation of points.}
}
\value{
A dataframe for input into pr_plot_Progress()
}
\description{
Prepare sampling coverage data for IMOS progress map visualisation
}
\examples{
df <- pr_get_ProgressMapData(c("NRS", "CPR"))
df <- pr_get_ProgressMapData(c("NRS", "CPR"), interactive = TRUE)
}
