% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pr_get_s3site}
\alias{pr_get_s3site}
\title{Get location of s3 plankton data}
\usage{
pr_get_s3site()
}
\value{
A string with location of s3 plankton data
}
\description{
Internal function to load the location of the s3 plankton data files.
}
\examples{
file_loc <- pr_get_s3site()
}
