% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sat.R
\name{pr_get_SatData}
\alias{pr_get_SatData}
\title{Load satellite-derived environmental data matched to plankton sampling locations}
\usage{
pr_get_SatData(Survey = "NRS")
}
\arguments{
\item{Survey}{either NRS or CPR}
}
\value{
dat with either NRS or CPR satellite data
}
\description{
Load satellite-derived environmental data matched to plankton sampling locations
}
\examples{
dat <- pr_get_SatData("CPR")
}
