% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pr_get_Site}
\alias{pr_get_Site}
\title{Get location of raw plankton data}
\usage{
pr_get_Site()
}
\value{
A string with location of raw plankton data
}
\description{
Internal function to load the location of the raw plankton data files.
}
\examples{
file_loc <- pr_get_Site()
}
