% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pr_get_site2}
\alias{pr_get_site2}
\title{Get location of raw plankton data on GitHub}
\usage{
pr_get_site2()
}
\value{
A string with location of raw plankton data
}
\description{
Internal function to load the location of the raw plankton data files stored on GitHub.
}
\examples{
file_loc <- pr_get_site2()
}
