% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sots.R
\name{pr_get_SOTSMoorData}
\alias{pr_get_SOTSMoorData}
\title{Get data available from SOTS moorings (limited to use in BOO atm, other variables could be added, select from pr_get_SOTSvariables)}
\usage{
pr_get_SOTSMoorData(Type = "Physical")
}
\arguments{
\item{Type}{Nutrients or Physical}
}
\value{
SOTS nutrients or physical parameters data
}
\description{
Get data available from SOTS moorings (limited to use in BOO atm, other variables could be added, select from pr_get_SOTSvariables)
}
\examples{
dat <- pr_get_SOTSMoorData(Type = "Nutrients")
}
