% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sots.R
\name{pr_get_SOTSvariables}
\alias{pr_get_SOTSvariables}
\title{Get data files available from SOTS moorings}
\usage{
pr_get_SOTSvariables(Type = "Physical")
}
\arguments{
\item{Type}{Physical or nutrients}
}
\value{
List of variables available from SOTS files
}
\description{
Get data files available from SOTS moorings
}
\examples{
df <- pr_get_SOTSvariables(Type = 'Physical')
}
