% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pr_get_StationName}
\alias{pr_get_StationName}
\title{Add StationName to data}
\usage{
pr_get_StationName(df)
}
\arguments{
\item{df}{A dataframe that contains \code{StationCode} but no \code{StationName}}
}
\value{
A dataframe with StationName added
}
\description{
Add StationName to data
}
\examples{
df <- pr_get_NRSStation() \%>\%
    pr_get_StationName()
}
