% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nrs.R
\name{pr_get_Stations}
\alias{pr_get_Stations}
\title{Import NRS Station information}
\usage{
pr_get_Stations(Survey = "NRS")
}
\arguments{
\item{Survey}{Survey type: "NRS" or "SOTS"}
}
\value{
A dataframe with station information
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{pr_get_Stations()} was deprecated in planktonr 0.7.0. Station information
is now available through \code{\link[=pr_get_info]{pr_get_info()}} with \code{Source = "NRS"} or
\code{Source = "SOTS"}.
}
\examples{
\dontrun{
# Instead of:
dat <- pr_get_Stations('NRS')

# Use:
dat <- pr_get_info(Source = "NRS")
}

}
\seealso{
\code{\link[=pr_get_info]{pr_get_info()}} for the preferred interface
}
\keyword{internal}
