% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_boo_data.R
\name{pr_get_STI}
\alias{pr_get_STI}
\title{Get data for STI plots of species abundance}
\usage{
pr_get_STI(Type = "P")
}
\arguments{
\item{Type}{Phyto or zoo, defaults to phyto}
}
\value{
df to be sued with pr_plot_STI
}
\description{
Get data for STI plots of species abundance
}
\examples{
df <- pr_get_STI("P")
df <- pr_get_STI("Z")
}
