% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planktonr_dat.R
\name{pr_get_survey.planktonr_dat}
\alias{pr_get_survey.planktonr_dat}
\title{Generic function and method for Survey}
\usage{
\method{pr_get_survey}{planktonr_dat}(x)
}
\arguments{
\item{x}{planktonr r data object}
}
\value{
attribute \code{Survey}
}
\description{
Generic function and method for Survey
}
