% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planktonr_dat.R
\name{pr_get_survey}
\alias{pr_get_survey}
\alias{pr_get_survey.planktonr_dat}
\title{Generic function to get the Survey attribute}
\usage{
pr_get_survey(x)

\method{pr_get_survey}{planktonr_dat}(x)
}
\arguments{
\item{x}{A \code{planktonr_dat} object.}
}
\value{
The \code{Survey} attribute from the object.
}
\description{
Generic function to get the Survey attribute
}
