% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_boo_data.R
\name{pr_get_TaxaAccum}
\alias{pr_get_TaxaAccum}
\title{Get taxa accumulation data for plotting}
\usage{
pr_get_TaxaAccum(Survey = "NRS", Type = "Z")
}
\arguments{
\item{Survey}{"NRS or "CPR"}

\item{Type}{"Z" or "P"}
}
\value{
A data frame
}
\description{
Get taxa accumulation data for plotting
}
\examples{
df <- pr_get_TaxaAccum(Survey = "NRS", Type = "Z")
}
