% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planktonr_dat.R
\name{pr_get_type}
\alias{pr_get_type}
\alias{pr_get_type.planktonr_dat}
\title{Generic function to get the Type attribute}
\usage{
pr_get_type(x)

\method{pr_get_type}{planktonr_dat}(x)
}
\arguments{
\item{x}{A \code{planktonr_dat} object.}
}
\value{
The \code{Type} attribute from the object.
}
\description{
Generic function to get the Type attribute
}
