% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planktonr_dat.R
\name{pr_get_variable}
\alias{pr_get_variable}
\alias{pr_get_variable.planktonr_dat}
\title{Generic function to get the Variable attribute}
\usage{
pr_get_variable(x)

\method{pr_get_variable}{planktonr_dat}(x)
}
\arguments{
\item{x}{A \code{planktonr_dat} object.}
}
\value{
The \code{Variable} attribute from the object.
}
\description{
Generic function to get the Variable attribute
}
