% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pr_get_ZooInfo}
\alias{pr_get_ZooInfo}
\title{Load copepod information table with sizes etc.}
\usage{
pr_get_ZooInfo()
}
\value{
A dataframe with NRS zooplankton information
}
\description{
Load copepod information table with sizes etc.
}
\examples{
df <- pr_get_ZooInfo()
}
