% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pr_harmonic}
\alias{pr_harmonic}
\title{For use in models to create a circular predictor}
\usage{
pr_harmonic(theta, k = 4)
}
\arguments{
\item{theta}{Paramter to model in radians}

\item{k}{degrees of freedom}
}
\value{
A harmonic function
}
\description{
This function is for use in models to create a circular predictor
}
\examples{
theta = 2
k = 1
df <- pr_harmonic(theta = 2, k = 1)
}
