% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_boo_plot.R
\name{pr_make_climatology}
\alias{pr_make_climatology}
\title{To produce the climatology for plotting}
\usage{
pr_make_climatology(df, x)
}
\arguments{
\item{df}{data frame containing columns Year, Month, Day}

\item{x}{Year, Month, Day, time period of climatology}
}
\value{
a dataframe for plotting
}
\description{
To produce the climatology for plotting
}
\examples{
df <- data.frame(Month = rep(1:12,10), Code = 'NSI', Values = runif(120, min=0, max=10))
pr_make_climatology(df, Month)
}
