% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sat.R
\name{pr_match_MODIS}
\alias{pr_match_MODIS}
\title{Match data for MODIS}
\usage{
pr_match_MODIS(df, pr)
}
\arguments{
\item{df}{dataframe containing latitude, longitude and Date}

\item{pr}{products from K_490, chl_carder, chl_gsm, chl_oc3, chl_oci, dt, ipar, l2_flags, owtd, par, sst, sst_quality single or as a list}
}
\value{
df with product output attached
}
\description{
Match data for MODIS
}
\examples{
df <- head(pr_get_DataLocs("NRS"),5)
res_spat = 10
MODISout <- pr_match_MODIS(df, pr <- c("chl_gsm", "chl_oc3"))
}
