% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planktonr_dat.R
\name{pr_planktonr_class}
\alias{pr_planktonr_class}
\title{Define the constructor function for the planktonr data class}
\usage{
pr_planktonr_class(
  data,
  type = NULL,
  survey = NULL,
  variable = NULL,
  subset = NULL
)
}
\arguments{
\item{data}{The data tibble to be updated to the planktonr class}

\item{type}{The data type - Microbes, Phytoplankton or Zooplankton}

\item{survey}{The survey of the data - NRS, CPR, LTM}

\item{variable}{What variable is being subsetted}

\item{subset}{NOT USED AT THE MOMENT}
}
\value{
An object of class planktonr_dat
}
\description{
Define the constructor function for the planktonr data class
}
