% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_boo_plot.R
\name{pr_plot_box}
\alias{pr_plot_box}
\title{Simple boxplot function using common NRS colouring}
\usage{
pr_plot_box(df, y)
}
\arguments{
\item{df}{Dataframe}

\item{y}{Column name for the y axis}
}
\value{
ggplot object
}
\description{
Note that this function assumes wide data with the data to plot as columns.
}
\examples{
df <- planktonr::pr_get_NRSMicro('Coastal') \%>\% tidyr::drop_na(tidyselect::all_of(c("Values", "Parameters"))) \%>\%
dplyr::filter(.data$StationCode \%in\% c("DEE", "DEB")) \%>\%
tidyr::pivot_wider(names_from = "Parameters", values_from = "Values", values_fn = 'mean')
gg <- pr_plot_box(df, "Bacterial_Temperature_Index_KD")
}
