% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_maps.R
\name{pr_plot_CPRmap}
\alias{pr_plot_CPRmap}
\title{Title Sidebar panel plot of selected CPR bioregions}
\usage{
pr_plot_CPRmap(sites)
}
\arguments{
\item{sites}{A string vector containing CPR bioregions to plot}
}
\value{
a map of the selected bioregions
}
\description{
Title Sidebar panel plot of selected CPR bioregions
}
\examples{
cprmap <- pr_plot_CPRmap(sites = c("Temperate East", "South-west",
                                    "South-east", "North", "North-west"))
}
