% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_boo_plot.R
\name{pr_plot_DayNight}
\alias{pr_plot_DayNight}
\title{Plot of relative day and night abundances}
\usage{
pr_plot_DayNight(df)
}
\arguments{
\item{df}{dataframe as output of pr_get_DayNight() filtered for one species}
}
\value{
plot of relative day and night abundances
}
\description{
Plot of relative day and night abundances
}
\examples{
df <- data.frame(Month = rep(seq(1,12,1),2), daynight = c(rep('day', 12), rep('night', 12)),
CopeAbundance_m3 = runif(24, 0.1, 10), Species = 'Acartia danae')
plot <- pr_plot_DayNight(df)
}
