% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_enviro.R
\name{pr_plot_Enviro}
\alias{pr_plot_Enviro}
\title{Combined timeseries and climatology plots for environmental variables}
\usage{
pr_plot_Enviro(df, Trend = "None", trans = "identity")
}
\arguments{
\item{df}{A dataframe from pr_get_nuts or pr_get_pigments}

\item{Trend}{Trend line to be used, options None, Smoother, Linear}

\item{trans}{scale on which to plot y axis}
}
\value{
A plot with timeseries and climatology at depths
}
\description{
Combined timeseries and climatology plots for environmental variables
}
\examples{
df <- pr_get_NRSChemistry() \%>\% dplyr::filter(
  Parameters == "SecchiDepth_m",
  StationCode \%in\% c("PHB", "NSI")
)
pr_plot_Enviro(df)

}
