% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_boo_plot.R
\name{pr_plot_EOV}
\alias{pr_plot_EOV}
\title{Policy plot}
\usage{
pr_plot_EOV(
  df,
  EOV = "Biomass_mgm3",
  Survey = "NRS",
  trans = "identity",
  pal = "matter",
  labels = "yes"
)
}
\arguments{
\item{df}{dataframe containing timeseries data with parameters and Values, output of pr_get_pol and pr_get_coeffs}

\item{EOV}{Essential OCean Variable as a parameter}

\item{Survey}{NRS, CPR or LTM 'Long term monitoring'}

\item{trans}{scale for y axis}

\item{pal}{colour pallet from cmocean}

\item{labels}{do you want to print labels on the x axes}
}
\value{
plot of timeseries, anomalies and climatology
}
\description{
Policy plot
}
\examples{
pr_get_pol("CPR") \%>\%
  pr_get_coeffs() \%>\%
  pr_plot_EOV(EOV = "Biomass_mgm3", Survey = "CPR",
      trans = "identity", pal = "matter", labels = "no")
}
