% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_boo_plot.R
\name{pr_plot_fmap}
\alias{pr_plot_fmap}
\title{Frequency plot of the selected species}
\usage{
pr_plot_fmap(df)
}
\arguments{
\item{df}{dataframe of format similar to output of pr_get_fmap_data()}
}
\value{
a plot of frequency of occurence of chosen species
}
\description{
Frequency plot of the selected species
}
\examples{
df <- data.frame(Long = c(110, 130, 155, 150), Lat = c(-10, -35, -27, -45),
freqfac = c("Absent", "Seen in 25\%",'50\%', '75\%'),
Season = c("December - February","March - May","June - August","September - November"),
Taxon = 'Acartia danae')
plot <- pr_plot_fmap(df)
}
