% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_maps.R
\name{pr_plot_FreqMap}
\alias{pr_plot_FreqMap}
\title{Map seasonal occurrence frequency for individual plankton species}
\usage{
pr_plot_FreqMap(dat, species, interactive = TRUE)
}
\arguments{
\item{dat}{dataframe of format similar to output of pr_get_fmap_data()}

\item{species}{species to plot}

\item{interactive}{ggplot if false, plotlist of leaflets if true}
}
\value{
a plot of frequency of occurrence of chosen species
}
\description{
Map seasonal occurrence frequency for individual plankton species
}
\examples{
dat <- data.frame(Longitude = c(110, 130, 155, 150), Latitude = c(-10, -35, -27, -45),
                 freqfac = as.factor(c("Absent", "Seen in 25\%",'50\%', '75\%')),
                 Season = c("December - February","March - May",
                 "June - August","September - November"),
                 Taxon = 'Acartia danae',
                 Survey = 'CPR')
plot <- pr_plot_FreqMap(dat, species = 'Acartia danae', interactive = TRUE)
}
