% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_info.R
\name{pr_plot_Gantt}
\alias{pr_plot_Gantt}
\title{Plot Gantt Chart showing plankton sampling status}
\usage{
pr_plot_Gantt(dat)
}
\arguments{
\item{dat}{Trip data for either NRS or CPR.}
}
\value{
a ggplot
}
\description{
Plot Gantt Chart showing plankton sampling status
}
\examples{
dat <- pr_get_CPRTrips()
gg <- pr_plot_Gantt(dat)
dat <- pr_get_NRSTrips()
gg <- pr_plot_Gantt(dat)
}
