% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_larvalfish.R
\name{pr_plot_LarvalFishDist}
\alias{pr_plot_LarvalFishDist}
\title{Plot of larval fish distribution}
\usage{
pr_plot_LarvalFishDist(df, SpeciesName, interactive = TRUE)
}
\arguments{
\item{df}{Dataframe of larval fish data from pr_get_LFData}

\item{SpeciesName}{The species you wish to plot}

\item{interactive}{Whether to return an interactive (leaflet; default) or ggplot.}
}
\value{
A plot element
}
\description{
Plot of larval fish distribution
}
\examples{
pr_get_LFData() \%>\%
   pr_plot_LarvalFishDist(SpeciesName = "Acanthuridae_37437900", interactive = TRUE)
}
