% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries.R
\name{pr_plot_latitude}
\alias{pr_plot_latitude}
\title{Create Hovmöller diagram showing latitudinal patterns over time}
\usage{
pr_plot_latitude(df, na.fill = TRUE)
}
\arguments{
\item{df}{dataframe of latitudinal series}

\item{na.fill}{TRUE, FALSE or function like mean to fill in gaps in data}
}
\value{
patchwork object
}
\description{
Create Hovmöller diagram showing latitudinal patterns over time
}
\examples{
df <- pr_get_data(Survey = 'GO-SHIP', Type = 'Micro')
df <- df \%>\% dplyr::filter(Parameters == 'Bacteria_unique_ASVs',
SampleDepth_m < 101)
pr_plot_latitude(df, na.fill = mean)
}
