% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_boo_plot.R
\name{pr_plot_NRSEnvContour}
\alias{pr_plot_NRSEnvContour}
\title{Contour plots for depth stratified environmental data}
\usage{
pr_plot_NRSEnvContour(df, Interpolation = TRUE)
}
\arguments{
\item{df}{dataframe from pr_get_NRSEnvContour}

\item{Interpolation}{If TRUE data is interpolated at 5m intervals (DAR = 10m)}
}
\value{
a contour plot
}
\description{
Contour plots for depth stratified environmental data
}
\examples{
df <- pr_get_NRSEnvContour("Pico") \%>\% dplyr::filter(Parameters == "Prochlorococcus_cellsmL",
StationCode \%in\% c('YON', 'PHB', 'NSI'))
plot <- pr_plot_NRSEnvContour(df, Interpolation = TRUE)
}
