% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_maps.R
\name{pr_plot_NRSmap}
\alias{pr_plot_NRSmap}
\title{Sidebar panel plot of selected NRS stations}
\usage{
pr_plot_NRSmap(sites)
}
\arguments{
\item{sites}{A string vector containing site codes to plot}
}
\value{
a map of the selected stations
}
\description{
Sidebar panel plot of selected NRS stations
}
\examples{
sites <- c("MAI", "PHB")
pmap <- pr_plot_NRSmap(sites)
}
