% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_boo_plot.R
\name{pr_plot_PCI}
\alias{pr_plot_PCI}
\title{PCI plot for CPR data}
\usage{
pr_plot_PCI(df)
}
\arguments{
\item{df}{dataframe with location and seasonal PCI data}
}
\value{
plot of PCI around Australia
}
\description{
PCI plot for CPR data
}
\examples{
pr_get_PCIData() \%>\% pr_plot_PCI()
}
