% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_info.R
\name{pr_plot_PieFG}
\alias{pr_plot_PieFG}
\title{Pie plots of functional groups for data}
\usage{
pr_plot_PieFG(df)
}
\arguments{
\item{df}{data frame binned to functional group e.g. planktonr::pr_get_FuncGroups("NRS", "Z")}
}
\value{
pie plot of functional groups
}
\description{
Pie plots of functional groups for data
}
\examples{
df <- pr_get_FuncGroups("CPR", "P")
plot <- pr_plot_PieFG(df)

}
