% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_boo_plot.R
\name{pr_plot_Progress}
\alias{pr_plot_Progress}
\title{IMOS progress plot}
\usage{
pr_plot_Progress(df)
}
\arguments{
\item{df}{output from pr_get_ProgressMap}
}
\value{
a plot of IMOS progress
}
\description{
IMOS progress plot
}
\examples{
df <- pr_get_ProgressMap("CPR")
plot <- pr_plot_Progress(df)
}
