% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_boo_plot.R
\name{pr_plot_ProgressMap}
\alias{pr_plot_ProgressMap}
\title{IMOS progress plot}
\usage{
pr_plot_ProgressMap(df, interactive = FALSE, labels = TRUE)
}
\arguments{
\item{df}{output from pr_get_ProgressMapData}

\item{interactive}{Should the plot be interactive with leaflet?}

\item{labels}{TRUE/FALSE Should labels be added to leaflet plot? Adding labels adds more information but slows down the rendering.}
}
\value{
a plot of IMOS progress
}
\description{
IMOS progress plot
}
\examples{
df <- pr_get_ProgressMapData("CPR")
plot <- pr_plot_ProgressMap(df)
}
