% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_boo_plot.R
\name{pr_plot_scatter}
\alias{pr_plot_scatter}
\title{Simple function to scatter 2 data columns using common NRS colouring}
\usage{
pr_plot_scatter(df, x, y)
}
\arguments{
\item{df}{Dataframe}

\item{x}{Column name for the x axis}

\item{y}{Column name for the y axis}
}
\value{
ggplot object
}
\description{
Note that this function assumes wide data with the data to plot as columns.
}
