% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_boo_plot.R
\name{pr_plot_STI}
\alias{pr_plot_STI}
\title{Plot of STI kernel density for species}
\usage{
pr_plot_STI(df)
}
\arguments{
\item{df}{dataframe as output of pr_get_STI() filtered for one species}
}
\value{
plot of STI kernel density
}
\description{
Plot of STI kernel density for species
}
\examples{
df <- data.frame(SST = runif(24, 5, 25),
                 Project = c(rep("CPR", 12), rep("NRS", 12)),
                 Species_m3 = runif(24, 0.1, 10),
                 Species = 'Acartia danae')

plot <- pr_plot_STI(df)
}
