% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_boo_plot.R
\name{pr_plot_TaxaAccum}
\alias{pr_plot_TaxaAccum}
\title{Taxa Accumulation Curve}
\usage{
pr_plot_TaxaAccum(dat, Survey = "NRS", Type = "Z")
}
\arguments{
\item{dat}{A dataframe of plankton data}

\item{Survey}{"CPR" or "NRS"}

\item{Type}{"P" or "Z"}
}
\value{
a ggplot object.
}
\description{
Plot a taxa accumulation curve for everything that is identified by the IMOS plankton team
}
\examples{
dat <- pr_get_TaxaAccum(Survey = "NRS", Type = "Z")
p <- pr_plot_TaxaAccum(dat, Survey = "NRS", Type = "Z")
dat <- pr_get_TaxaAccum(Survey = "CPR", Type = "P")
p <- pr_plot_TaxaAccum(dat, Survey = "CPR", Type = "P")
}
