% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries.R
\name{pr_plot_TimeSeries}
\alias{pr_plot_TimeSeries}
\title{Plot basic timeseries}
\usage{
pr_plot_TimeSeries(df, trans = "identity")
}
\arguments{
\item{df}{dataframe with SampleDate_Local, station code and parameter name and values}

\item{trans}{scale of y axis on plot, whatever scale_y_continuous trans accepts}
}
\value{
a timeseries plot
}
\description{
Plot basic timeseries
}
\examples{
df <- pr_get_Indices("NRS", "Zooplankton") \%>\%
  dplyr::filter(Parameters == "Biomass_mgm3", StationCode \%in\% c("NSI", "PHB"))
timeseries <- pr_plot_TimeSeries(df)
}
