% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries.R
\name{pr_plot_tsclimate}
\alias{pr_plot_tsclimate}
\title{Combined timeseries and climatology plots}
\usage{
pr_plot_tsclimate(df, trans = "identity")
}
\arguments{
\item{df}{data frame with SampleDate_Local, time period and parameter}

\item{trans}{scale of y axis on plot, whatever scale_y_continuous trans accepts}
}
\value{
a combined plot
}
\description{
Combined timeseries and climatology plots
}
\examples{
df <- pr_get_Indices(Survey = "CPR", Type = "Phytoplankton") \%>\%
  dplyr::filter(Parameters == "PhytoAbundance_Cellsm3")
pr_plot_tsclimate(df)
}
