% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_boo_plot.R
\name{pr_plot_tsfg}
\alias{pr_plot_tsfg}
\title{Time series plot of functional groups}
\usage{
pr_plot_tsfg(df, Scale = "Actual", trend = "Raw")
}
\arguments{
\item{df}{dataframe in format of output from pr_get_fg}

\item{Scale}{y axis scale Actual or Percent}

\item{trend}{Over what timescale to fit the trend - "Raw", "Month" or "Year"}
}
\value{
plot of fg timseries
}
\description{
Time series plot of functional groups
}
\examples{
df <- pr_get_fg("NRS", "P")
plot <- pr_plot_tsfg(df, "Actual")
}
