% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_maps.R
\name{pr_plot_Voyagemap}
\alias{pr_plot_Voyagemap}
\title{Create map showing CPR voyage tracks and sampling locations}
\usage{
pr_plot_Voyagemap(dat, dat_select, Country = c("Australia"))
}
\arguments{
\item{dat}{dataframe containing all locations to plot}

\item{dat_select}{dataframe of sample locations to highlight}

\item{Country}{countries to plot on map}
}
\value{
a map of the selected bioregions
}
\description{
Create map showing CPR voyage tracks and sampling locations
}
\examples{
dat <- pr_get_data(Survey = "GO-SHIP", Type = "Micro")
dat_select <- dat \%>\% dplyr::slice(1:5000)
voyagemap <- pr_plot_Voyagemap(dat, dat_select, Country = c("Australia", "New Zealand"))
}
