% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_relabel.R
\name{pr_relabel}
\alias{pr_relabel}
\title{Format a parameter string for plotting}
\usage{
pr_relabel(s, style = "ggplot", named = FALSE, quiet = FALSE)
}
\arguments{
\item{s}{A string for reformatting}

\item{style}{The style of plotting the string will be used for}

\item{named}{Should the output be an \code{unnamed} (FALSE) or \code{named} (TRUE) vector? Only available for \code{style = "simple"} and \code{style = "plotly"}}

\item{quiet}{Logical. If \code{TRUE}, suppresses warnings about missing replacements. Default is \code{FALSE}.}
}
\value{
A reformatted expression call (ggplot)
}
\description{
Format a parameter string for plotting
}
\keyword{internal}
