% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pr_remove_outliers}
\alias{pr_remove_outliers}
\title{Removing outliers from data}
\usage{
pr_remove_outliers(df, x)
}
\arguments{
\item{df}{data frame with Parameters, Values and StationCode or BioRegion}

\item{x}{no of sd from mean to remove}
}
\value{
df
}
\description{
Removing outliers from data
}
\examples{
df <- pr_get_Indices("NRS", "Z") \%>\% pr_remove_outliers(2)
df <- pr_get_Indices("CPR", "Z") \%>\% pr_remove_outliers(2)
}
