% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_rename.R
\name{pr_rename}
\alias{pr_rename}
\title{Rename variables within dataframe}
\usage{
pr_rename(df)
}
\arguments{
\item{df}{Dataframe with column names to renamed}
}
\value{
Dataframe with new column names
}
\description{
Rename variables within dataframe
}
