% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pr_reorder}
\alias{pr_reorder}
\title{Order the Station or region in the df by latitude for consistent plotting}
\usage{
pr_reorder(df)
}
\arguments{
\item{df}{A dataframe that contains StationName, StationCode or BioRegion}
}
\value{
An ordered dataframe
}
\description{
Order the Station or region in the df by latitude for consistent plotting
}
\examples{
df <- data.frame(StationName = c('Port Hacking', 'Maria Island',
'North Stradbroke Island','Esperance', 'Ningaloo', "Darwin",
'Rottnest Island',  'Kangaroo Island', "Yongala")) \%>\%
planktonr_dat(Survey = "NRS")
df <- pr_reorder(df)
}
