% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pr_title}
\alias{pr_title}
\title{Helper function to reformat Titles}
\usage{
pr_title(tit)
}
\arguments{
\item{tit}{String to reformat.}
}
\value{
A reformatted string
}
\description{
Helper function to reformat Titles
}
\examples{
new_tit = pr_title("Phytoplankton")
}
