% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{summarise.planktonr_dat}
\alias{summarise.planktonr_dat}
\title{Summarise data in a planktonr_dat object}
\usage{
\method{summarise}{planktonr_dat}(.data, ...)
}
\arguments{
\item{.data}{A \code{planktonr_dat} object.}

\item{...}{<\code{\link[dplyr:dplyr_data_masking]{data-masking}}> Name-value pairs of summary statistics,
and optional arguments \code{.by} and \code{.groups}. See \code{\link[dplyr:summarise]{dplyr::summarise()}} for details.}
}
\value{
A \code{planktonr_dat} object with summarised data,
preserving original attributes. The grouping structure depends on \code{.groups}.
}
\description{
This is an S3 method for \code{dplyr::summarise} that ensures
\code{planktonr_dat} attributes are preserved during summarisation.
}
