% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{untibble}
\alias{untibble}
\title{Remove extra tibble info}
\usage{
untibble(tib)
}
\arguments{
\item{tib}{}
}
\value{
A dataframe with with all the tibble stuff removed
}
\description{
Remove extra tibble info
}
\examples{
data(mtcars)
tib <- dplyr::as_tibble(mtcars)
df <- untibble(tib)
}
