% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planktonr_dat.R
\name{is_planktonr_dat}
\alias{is_planktonr_dat}
\title{Check if an object is of class \code{planktonr_dat}}
\usage{
is_planktonr_dat(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
\code{TRUE} if \code{x} is a \code{planktonr_dat} object, otherwise \code{FALSE}.
}
\description{
This function tests if the given object \code{x} inherits from the \code{planktonr_dat} class.
}
\examples{
# Assuming planktonr_dat constructor exists:
my_data <- planktonr_dat(tibble::tibble(a = 1:3), type = "Phytoplankton")
is_planktonr_dat(my_data) # Returns TRUE
is_planktonr_dat(tibble::tibble(a = 1:3)) # Returns FALSE
}
