% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planktonr_dat.R
\name{pivot_longer.planktonr_dat}
\alias{pivot_longer.planktonr_dat}
\title{pivot_longer method for planktonr_dat}
\usage{
pivot_longer.planktonr_dat(data, ...)
}
\arguments{
\item{data}{The planktonr data object}

\item{...}{Other possible objects to be passed to the tidyverse}
}
\description{
pivot_longer method for planktonr_dat
}
