% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planktonr_dat.R
\name{planktonr_dat}
\alias{planktonr_dat}
\title{Define the constructor function for the planktonr data class}
\usage{
planktonr_dat(
  .data,
  Type = NULL,
  Survey = NULL,
  Variable = NULL,
  Model = NULL,
  ...
)
}
\arguments{
\item{.data}{The data \code{data.frame} or \code{tibble} to be converted to the \code{planktonr_dat} class.}

\item{Type}{The data type. Must be one of "Microbes", "Phytoplankton",
"Zooplankton", "Water", or "EOV".}

\item{Survey}{The survey of the data. E.g., "NRS", "CPR", "LTM", "GO-SHIP", "Coastal".}

\item{Variable}{What variable is being described or subsetted by this data.}

\item{Model}{Optional model associated with the data.}

\item{...}{Additional attributes to be stored with the \code{planktonr_dat} object.
These will be stored as attributes on the object.}
}
\value{
An object of class \code{planktonr_dat}.
}
\description{
This function creates a \code{planktonr_dat} object, which is a tibble
with additional custom attributes for metadata. It can accept
either a \code{data.frame} or a \code{tibble}, coercing \code{data.frame} inputs to \code{tibble}.
}
\examples{
# Create from a tibble (existing behavior)
tibble_data <- tibble::tibble(
  time = as.Date(c("2023-01-01", "2023-01-02")),
  value = c(10, 20)
)
pr_obj_tibble <- planktonr_dat(
  .data = tibble_data,
  Type = "Phytoplankton",
  Survey = "NRS",
  Variable = "Chlorophyll_a"
)
print(pr_obj_tibble)

# Create from a base R data.frame (newly supported)
df_data <- data.frame(
  time = as.Date(c("2023-01-03", "2023-01-04")),
  value = c(30, 40)
)
pr_obj_df <- planktonr_dat(
  .data = df_data,
  Type = "Zooplankton",
  Survey = "CPR",
  Variable = "Biomass"
)
print(pr_obj_df)
class(pr_obj_df) # Still a planktonr_dat (which inherits from tibble)
}
