% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_add_bioregions.R
\name{pr_add_Bioregions}
\alias{pr_add_Bioregions}
\title{Add bioregions data to existing df}
\usage{
pr_add_Bioregions(df, join = "st_within", ...)
}
\arguments{
\item{df}{A dataframe with columns \code{.data$Longitude} and \code{.data$Latitude}}

\item{join}{use join arguments from sf::st_join for joining to CPR samples to bioregions.}

\item{...}{to allow use of join when used within another function}
}
\value{
A dataframe with Marine Bioregions added
}
\description{
Add bioregions data to existing df
}
\examples{
 df <- pr_get_Raw("cpr_derived_indices_data") \%>\%
  pr_rename() \%>\%
   pr_add_Bioregions("st_nearest_feature")
}
