% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pr_add_SampleDate}
\alias{pr_add_SampleDate}
\title{Add the SampleDate to dataframes that only have SampleTime}
\usage{
pr_add_SampleDate(df)
}
\arguments{
\item{df}{Dataframe containing \code{SampleTime_Local} or \code{SampleTime_UTC}}
}
\value{
A dataframe with \code{SampleDate} included.
}
\description{
Add the SampleDate to dataframes that only have SampleTime
}
\examples{
df <- pr_add_SampleDate(data.frame(SampleTime_Local = Sys.time()))
}
