% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pr_add_StationCode}
\alias{pr_add_StationCode}
\title{Add NRS StationCode to data}
\usage{
pr_add_StationCode(df)
}
\arguments{
\item{df}{A dataframe that contains \code{StationName} but no \code{StationCode}}
}
\value{
A dataframe with StationCode added
}
\description{
Add NRS StationCode to data
}
\examples{
df <- pr_get_NRSStation() \%>\%
    pr_add_StationCode()
}
