% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pr_apply_flags}
\alias{pr_apply_flags}
\title{Remove flagged data in df}
\usage{
pr_apply_flags(df)
}
\arguments{
\item{df}{A dataframe containing data with associated flags}
}
\value{
A dataframe with flagged data removed
}
\description{
Remove flagged data in df
}
\examples{
df <- data.frame(SST = c(27.4, 28.9, 45), SST_Flag = c(1, 1, 4))
df <- pr_apply_flags(df)
}
