% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pr_apply_time}
\alias{pr_apply_time}
\title{Remove flagged data in df}
\usage{
pr_apply_time(df)
}
\arguments{
\item{df}{A dataframe containing time column}
}
\value{
A dataframe with extra date columns
}
\description{
Remove flagged data in df
}
\examples{
df <- data.frame(SampleDate_Local = lubridate::now(), Latitude = -32, Longitude = 160)
df <- pr_apply_time(df)
}
