% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pr_apply_Time}
\alias{pr_apply_Time}
\title{Add Month & year columns to df}
\usage{
pr_apply_Time(df)
}
\arguments{
\item{df}{A dataframe containing time column}
}
\value{
A dataframe with extra date columns
}
\description{
Add Month & year columns to df
}
\examples{
df <- pr_get_Indices("NRS", "P") \%>\%
  pr_apply_Time()
}
