% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_indices.R
\name{pr_filter_data}
\alias{pr_filter_data}
\title{Filter data for plotting functions}
\usage{
pr_filter_data(df, Parameter = "Biomass_mgm3", StationRegion = "NSI")
}
\arguments{
\item{df}{dataframe from pr_get_indices}

\item{Parameter}{parameter(s) to be filtered on}

\item{StationRegion}{StationCode(s) or BioRegion(s) to be filtered on}
}
\value{
prefiltered data frame for pr_plot_trends, ts, climate
}
\description{
Filter data for plotting functions
}
\examples{
df <- pr_get_Indices("CPR", "Z") \%>\%
      pr_filter_data("BiomassIndex_mgm3", c("North", "South-west"))
df <- pr_get_Indices("NRS", "P") \%>\%
      pr_filter_data("PhytoBiomassCarbon_pgL", c("NSI", "PHB"))
}
