% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pr_get_coeffs}
\alias{pr_get_coeffs}
\title{Get coefficients from linear model}
\usage{
pr_get_coeffs(df)
}
\arguments{
\item{df}{dataframe containing Year, Month, parameters, values}
}
\value{
coefficients from linear model
}
\description{
Get coefficients from linear model
}
\examples{
df <- data.frame(Year_Local = runif(10, 2000, 2003),
                 Month_Local = runif(10, 1, 6),
                 parameters = c('Biomasss_mgm3', 'Diversity'),
                 Values = runif(10, 1, 5))
pr <- pr_get_coeffs(df)
}
