% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pr_get_coeffs}
\alias{pr_get_coeffs}
\title{Get coefficients from linear model}
\usage{
pr_get_coeffs(df)
}
\arguments{
\item{df}{dataframe containing Year, Month, parameters, values}
}
\value{
coefficients from linear model
}
\description{
Get coefficients from linear model
}
\examples{
df <- planktonr::pr_get_pol("NRS") \%>\%
  pr_get_coeffs()
}
