% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pr_get_Coeffs}
\alias{pr_get_Coeffs}
\title{Get coefficients from linear model}
\usage{
pr_get_Coeffs(df)
}
\arguments{
\item{df}{dataframe containing Year, Month, Parameters, values}
}
\value{
coefficients from linear model
}
\description{
Get coefficients from linear model
}
\examples{
df <- planktonr::pr_get_EOVs("NRS") \%>\%
  dplyr::filter(
    StationCode \%in\% c("PHB", "NSI"),
    Parameters \%in\% c("Biomass_mgm3", "PhytoBiomassCarbon_pgL")
  ) \%>\%
  pr_get_Coeffs()
}
