% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cpr.R
\name{pr_get_CPRData}
\alias{pr_get_CPRData}
\title{Import CPR Data}
\usage{
pr_get_CPRData(Type = "phytoplankton", Variable = "abundance", Subset = "raw")
}
\arguments{
\item{Type}{The data of interest: Phytoplankton or Zooplankton}

\item{Variable}{Variable options are: abundance or biovolume (phytoplankton only)}

\item{Subset}{Data compilation. Full options are below.}
}
\value{
A dataframe with requested plankton data in long form
}
\description{
Load CPR Data
}
\examples{
df <- pr_get_CPRData(Type = "phytoplankton", Variable = "abundance", Subset = "raw")
}
