% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cpr.R
\name{pr_get_CPRPhytoData}
\alias{pr_get_CPRPhytoData}
\title{Get CPR Phytoplankton Abundance data}
\usage{
pr_get_CPRPhytoData()
}
\value{
A dataframe with CPR Phytoplankton Abundance data
}
\description{
Get CPR Phytoplankton Abundance data
}
\examples{
df <- pr_get_CPRPhytoData()
}
