% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cpr.R
\name{pr_get_CPRSamps}
\alias{pr_get_CPRSamps}
\title{Get CPR samples}
\usage{
pr_get_CPRSamps(Type = c("P", "Z", "B"))
}
\arguments{
\item{Type}{A character string on which to filter data (P = Phytoplankton, Z = Zooplankton, B = Biomass)}
}
\value{
A dataframe with CPR Samples
}
\description{
Get CPR samples
}
\examples{
df <- pr_get_CPRSamps("Z")
}
