% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cpr.R
\name{pr_get_CPRZooData}
\alias{pr_get_CPRZooData}
\title{Get CPR Zooplankton abundance data}
\usage{
pr_get_CPRZooData(var = "Abundance")
}
\arguments{
\item{var}{A string of either "Abundance" or "Count"}
}
\value{
A dataframe with CPR Zooplankton Abundance data
}
\description{
Get CPR Zooplankton abundance data
}
\examples{
df <- pr_get_CPRZooData()
}
