% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_bgc.R
\name{pr_get_CSChem}
\alias{pr_get_CSChem}
\title{Load microbial data}
\usage{
pr_get_CSChem()
}
\arguments{
\item{Survey}{NRS or Coastal stations}
}
\value{
A dataframe with NRS microbial data
}
\description{
Load microbial data
}
\examples{
df <- pr_get_CSChem()
}
