% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_bgc.R
\name{pr_get_CSChem}
\alias{pr_get_CSChem}
\title{Load Coastal Seas chemistry and nutrient data}
\usage{
pr_get_CSChem()
}
\value{
A dataframe with NRS chemistry data
}
\description{
Load Coastal Seas chemistry and nutrient data
}
\examples{
df <- pr_get_CSChem()
}
