% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plankton.R
\name{pr_get_CTI}
\alias{pr_get_CTI}
\title{Get CTI for sample}
\usage{
pr_get_CTI(Type = "Z")
}
\arguments{
\item{Type}{Zooplankton (Copepods) or phytoplankton species}
}
\value{
df of CTI for each sample
}
\description{
Get CTI for sample
}
\examples{
df <- pr_get_CTI("P")

}
