% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sat.R
\name{pr_get_DataLocs}
\alias{pr_get_DataLocs}
\title{Functions for matching location data to satellite products}
\usage{
pr_get_DataLocs(Survey = "all")
}
\arguments{
\item{Survey}{NRS, CPR or all}
}
\value{
df with latitude, longitude and Date
}
\description{
Get data for satellite matching
}
\examples{
df <- pr_get_DataLocs("NRS")
}
