% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_boo_data.R
\name{pr_get_daynight}
\alias{pr_get_daynight}
\title{Get data for plots of species abundance by day and night using CPR data}
\usage{
pr_get_daynight(Type = "Z")
}
\arguments{
\item{Type}{Phyto or zoo, defaults to phyto}
}
\value{
df to be sued with pr_plot_daynight
}
\description{
Get data for plots of species abundance by day and night using CPR data
}
\examples{
df <- pr_get_daynight("Z")
}
