% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_policy.R
\name{pr_get_EOVs}
\alias{pr_get_EOVs}
\title{Load Essential Ocean Variables (EOVs) for plankton biomass and diversity}
\usage{
pr_get_EOVs(Survey = "NRS", ...)
}
\arguments{
\item{Survey}{"NRS" or "CPR" or "LTM"}

\item{...}{to allow use of join when used within another function}
}
\value{
A dataframe with policy data
}
\description{
Load Essential Ocean Variables (EOVs) for plankton biomass and diversity
}
\examples{
df <- pr_get_EOVs("NRS")
}
