% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_indices.R
\name{pr_get_indices}
\alias{pr_get_indices}
\title{Access data for timeseries and climatology plots}
\usage{
pr_get_indices(Survey = "CPR", Type = "P")
}
\arguments{
\item{Survey}{CPR or NRS, defaults to NRS}

\item{Type}{Phyto or zoo, defaults to phyto}
}
\value{
dataframe to use in pr_plot functions
}
\description{
Access data for timeseries and climatology plots
}
\examples{
df <- pr_get_indices("NRS", "P")
df <- pr_get_indices("NRS", "Z")
df <- pr_get_indices("CPR", "P")
df <- pr_get_indices("CPR", "Z")
}
