% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_indices.R
\name{pr_get_Indices}
\alias{pr_get_Indices}
\title{Access data for timeseries and climatology plots}
\usage{
pr_get_Indices(Survey = "CPR", Type = "P", ...)
}
\arguments{
\item{Survey}{CPR or NRS, defaults to NRS}

\item{Type}{Phyto, zoo or physical water props, defaults to phyto}

\item{...}{variables to be passed to pr_add_Bioregions. At the moment it only supports \code{near_dist_km} which is the distance (km) around each bioregion to pad the allocation of points.}
}
\value{
dataframe to use in pr_plot functions
}
\description{
Access data for timeseries and climatology plots
}
\examples{
df <- pr_get_Indices("NRS", "P")
df <- pr_get_Indices("NRS", "Z")
df <- pr_get_Indices("NRS", "W")
df <- pr_get_Indices("CPR", "P", near_dist_km = 250)
df <- pr_get_Indices("CPR", "Z")
}
