% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_larvalfish.R
\name{pr_get_LFData}
\alias{pr_get_LFData}
\title{Get Larval Fish Sample Data}
\usage{
pr_get_LFData()
}
\value{

}
\description{
Get Larval Fish Sample Data
}
\examples{
df <- pr_get_LFData()
#' @importFrom magrittr "\%>\%"
}
