% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_bgc.R
\name{pr_get_LTnuts}
\alias{pr_get_LTnuts}
\title{Get NRS long term nutrient timeseries data}
\usage{
pr_get_LTnuts()
}
\value{
dataframe for plotting long term nutrient time series info
}
\description{
Get NRS long term nutrient timeseries data
}
\examples{
dat <- pr_get_LTnuts()
}
