% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planktonr_dat.R
\name{pr_get_model}
\alias{pr_get_model}
\title{Generic function and method for Model}
\usage{
pr_get_model(x)
}
\arguments{
\item{x}{planktonr r data object}
}
\value{
attribute \code{Model}
}
\description{
Generic function and method for Model
}
