% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pr_get_NonTaxaColumns}
\alias{pr_get_NonTaxaColumns}
\title{Return columns that are not taxonomic}
\usage{
pr_get_NonTaxaColumns(Survey = "NRS", Type = "Zooplankton")
}
\arguments{
\item{Survey}{\code{NRS} or \code{CPR}}

\item{Type}{\code{zooplankton} or \code{phytoplankton}}
}
\value{
a string vector of column name
}
\description{
Return columns that are not taxonomic
}
\keyword{internal}
