% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_get_bgc.R
\name{pr_get_NRSChemistry}
\alias{pr_get_NRSChemistry}
\title{Load chemistry data}
\usage{
pr_get_NRSChemistry()
}
\value{
A dataframe with NRS chemistry data
}
\description{
Load chemistry data
}
\examples{
df <- pr_get_NRSChemistry()
}
