% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_bgc.R
\name{pr_get_NRSCTD}
\alias{pr_get_NRSCTD}
\title{Load NRS CTD profile data (temperature, salinity, fluorescence)}
\usage{
pr_get_NRSCTD()
}
\value{
A dataframe with NRS CTD data
}
\description{
Load NRS CTD profile data (temperature, salinity, fluorescence)
}
\examples{
dat <- pr_get_NRSCTD()
}
