% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nrs.R
\name{pr_get_NRSData}
\alias{pr_get_NRSData}
\title{Import NRS Data}
\usage{
pr_get_NRSData(Type = "phytoplankton", Variable = "abundance", Subset = "raw")
}
\arguments{
\item{Type}{The data of interest: Phytoplankton or Zooplankton}

\item{Variable}{Variable options are: abundance or biovolume (phytoplankton only)}

\item{Subset}{Data compilation. Full options are below.}
}
\value{
A dataframe with requested plankton data in wide form
}
\description{
Load NRS station Data
}
\details{
nrs_phytoplankton_abundance_raw_data
nrs_phytoplankton_abundance_htg_data
nrs_phytoplankton_abundance_genus_data
nrs_phytoplankton_abundance_species_data

nrs_phytoplankton_biovolume_raw_data
nrs_phytoplankton_biovolume_htg_data
nrs_phytoplankton_biovolume_genus_data
nrs_phytoplankton_biovolume_species_data

nrs_zooplankton_abundance_copepods_data
nrs_zooplankton_abundance_non_copepods_data
nrs_zooplankton_abundance_genus_data
nrs_zooplankton_abundance_htg_data
nrs_zooplankton_abundance_raw_data
}
\examples{
df <- pr_get_NRSData(Type = "phytoplankton", Variable = "abundance", Subset = "raw")
}
