% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nrs.R
\name{pr_get_NRSData}
\alias{pr_get_NRSData}
\title{Import NRS Data}
\usage{
pr_get_NRSData(file)
}
\arguments{
\item{file}{Filename to retrieve from AODN.}
}
\value{
A dataframe with requested plankton data in long form
}
\description{
Load NRS station Data
}
\examples{
df <- pr_get_NRSData("bgc_phytoplankton_abundance_htg_data")
}
