% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{pr_get_NRSData}
\alias{pr_get_NRSData}
\title{Import NRS plankton data}
\usage{
pr_get_NRSData(Type = "Phytoplankton", Variable = "abundance", Subset = "raw")
}
\arguments{
\item{Type}{The data of interest: \code{"Phytoplankton"} or \code{"Zooplankton"}}

\item{Variable}{Variable: \code{"abundance"} or \code{"biovolume"}}

\item{Subset}{Data level: \code{"raw"}, \code{"htg"}, \code{"genus"}, \code{"species"}, or \code{"copepods"}}
}
\value{
A dataframe with NRS plankton data.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{pr_get_NRSData()} was deprecated in planktonr 0.7.0 in favour of
\code{\link[=pr_get_data]{pr_get_data()}} which provides a unified interface for all data types.
}
\examples{
\dontrun{
# Instead of:
dat <- pr_get_NRSData(Type = "Phytoplankton", Variable = "abundance", Subset = "raw")

# Use:
dat <- pr_get_data(Survey = "NRS", Type = "Phytoplankton",
                   Variable = "abundance", Subset = "raw")
}
}
\seealso{
\code{\link[=pr_get_data]{pr_get_data()}} for the preferred interface
}
\keyword{internal}
