% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nrs.R
\name{pr_get_NRSData}
\alias{pr_get_NRSData}
\title{Import NRS Data}
\usage{
pr_get_NRSData(Type = "Phytoplankton", Variable = "abundance", Subset = "raw")
}
\arguments{
\item{Type}{The data of interest: Phytoplankton or Zooplankton}

\item{Variable}{Variable options are: abundance or biovolume (phytoplankton only)}

\item{Subset}{Data compilation. Full options are below.}
}
\value{
A dataframe with requested plankton data in wide form
}
\description{
Load NRS station Data
}
\examples{
df <- pr_get_NRSData(Type = "Phytoplankton", Variable = "abundance", Subset = "raw")

}
