% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nrs.R
\name{pr_get_NRSPhytoHTG}
\alias{pr_get_NRSPhytoHTG}
\title{Get Abundance of Phyto Higher Trophic Groups}
\usage{
pr_get_NRSPhytoHTG()
}
\value{
A dataframe with Phytoplankton Abundance data - Summed by Higher Trophic Levels
}
\description{
Get Abundance of Phyto Higher Trophic Groups
}
\examples{
df <- pr_get_NRSPhytoHTG()
#' @import dplyr
}
