% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_get_bgc.R
\name{pr_get_NRSPico}
\alias{pr_get_NRSPico}
\title{Load picophytoplankton data}
\usage{
pr_get_NRSPico()
}
\value{
A dataframe with NRS picophytoplankton data
}
\description{
Load picophytoplankton data
}
\examples{
df <- pr_get_NRSPico()
}
