% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_bgc.R
\name{pr_get_NRSPigments}
\alias{pr_get_NRSPigments}
\title{Get pigments data}
\usage{
pr_get_NRSPigments(format = "all")
}
\arguments{
\item{format}{all or binned}
}
\value{
A dataframe with NRS pigment data
}
\description{
Get pigments data
}
\examples{
df <- pr_get_NRSPigments()
df <- pr_get_NRSPigments('binned')
}
