% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_get_bgc.R
\name{pr_get_NRSPigments}
\alias{pr_get_NRSPigments}
\title{Get pigments data}
\usage{
pr_get_NRSPigments()
}
\value{
A dataframe with NRS pigment data
}
\description{
Get pigments data
}
\examples{
df <- pr_get_NRSPigments()
}
