% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nrs.R
\name{pr_get_NRSTrips}
\alias{pr_get_NRSTrips}
\title{Import NRS BGC information}
\usage{
pr_get_NRSTrips(Type = c("Phytoplankton", "Zooplankton", "Fish"))
}
\arguments{
\item{Type}{A character string on which to filter data (P = Phytoplankton, Z = Zooplankton, F = Fish)}
}
\value{
A dataframe with NRS BGC information
}
\description{
Load NRS station BGC information
}
\examples{
df <- pr_get_NRSTrips(Type = "Zooplankton")
}
