% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_get_bgc.R
\name{pr_get_NRSTSS}
\alias{pr_get_NRSTSS}
\title{Load Total Suspended Solids (TSS) data}
\usage{
pr_get_NRSTSS()
}
\value{
A dataframe with NRS TSS data
}
\description{
Load Total Suspended Solids (TSS) data
}
\examples{
df <- pr_get_NRSTSS()
}
