% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_boo_data.R
\name{pr_get_Papers}
\alias{pr_get_Papers}
\title{Random scientific papers using IMOS data}
\usage{
pr_get_Papers()
}
\value{
A random publication reference as a string
}
\description{
This function randomly returns a publication that uses the IMOS plankton data
}
\examples{
pr_get_Papers()
}
