% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plankton.R
\name{pr_get_PCIData}
\alias{pr_get_PCIData}
\title{Get Phytoplankton Colour Index (PCI) data from CPR samples}
\usage{
pr_get_PCIData()
}
\value{
dataframe of PCI data
}
\description{
Get Phytoplankton Colour Index (PCI) data from CPR samples
}
\examples{
head(pr_get_PCIData(),5)
}
