% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plankton.R
\name{pr_get_PCIData}
\alias{pr_get_PCIData}
\title{Data for PCI plot from CPR data}
\usage{
pr_get_PCIData()
}
\value{
dataframe of PCI data
}
\description{
Data for PCI plot from CPR data
}
\examples{
head(pr_get_PCIData(), 5)
}
