% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pr_get_PlanktonInfo}
\alias{pr_get_PlanktonInfo}
\title{Load plankton information table with sizes etc.}
\usage{
pr_get_PlanktonInfo(Type = "Zooplankton")
}
\arguments{
\item{Type}{Get info for Phytoplankton (\code{P}) or Zooplankton(\code{Z})}
}
\value{
A dataframe with NRS zooplankton information
}
\description{
Load plankton information table with sizes etc.
}
\examples{
df <- pr_get_PlanktonInfo(Type = "Phytoplankton")
df <- pr_get_PlanktonInfo(Type = "Zooplankton")
}
