% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_policy.R
\name{pr_get_PolicyInfo}
\alias{pr_get_PolicyInfo}
\title{Load policy-relevant plankton indicators and Essential Ocean Variables}
\usage{
pr_get_PolicyInfo(Survey = "NRS", ...)
}
\arguments{
\item{Survey}{"NRS" or "CPR"}

\item{...}{to allow use of join when used within another function}
}
\value{
A dataframe with policy information
}
\description{
Load policy-relevant plankton indicators and Essential Ocean Variables
}
\examples{
df <- pr_get_PolicyInfo("CPR")
}
