% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_boo_data.R
\name{pr_get_ProgressMapData}
\alias{pr_get_ProgressMapData}
\title{Data for IMOS progress map}
\usage{
pr_get_ProgressMapData(Survey = c("NRS", "CPR"), interactive = FALSE, ...)
}
\arguments{
\item{Survey}{one of NRS, CPR or Both}

\item{interactive}{A logical TRUE/FALSE if the data is to be used for an interactive plot.}
}
\value{
A dataframe for input into pr_plot_Progress()
}
\description{
Data for IMOS progress map
}
\examples{
df <- pr_get_ProgressMapData(c("NRS", "CPR"))
df <- pr_get_ProgressMapData(c("NRS", "CPR"), interactive = TRUE)
}
