% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pr_get_s3}
\alias{pr_get_s3}
\title{Download Raw Data from IMOS}
\usage{
pr_get_s3(file)
}
\arguments{
\item{file}{The filename of the requested data}
}
\value{
A dataframe of raw IMOS data
}
\description{
This function is not intended for use by most people. It downloads the raw data file from IMOS with NO QC or data manipulation done. User beware. It can be used to view the raw data that IMOS provides behind the scenes.
}
\details{
Options for \code{file} include: "bgc_trip"
}
\examples{
dat <- pr_get_s3("bgc_trip")
}
