% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sat.R
\name{pr_get_SatData}
\alias{pr_get_SatData}
\title{Get data for satellite data}
\usage{
pr_get_SatData(Survey = "NRS")
}
\arguments{
\item{Survey}{either NRS or CPR}
}
\value{
df with either NRS or CPR satellite data
}
\description{
Get data for satellite data
}
\examples{
df <- pr_get_SatData("CPR")
}
