% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pr_get_SpeciesInfo}
\alias{pr_get_SpeciesInfo}
\title{Get species information table for Phytoplankton and Zooplankton}
\usage{
pr_get_SpeciesInfo(Type = "Zooplankton")
}
\arguments{
\item{Type}{Phytoplankton (P) or Zooplankton (Z)}
}
\value{
A dataframe of species information
}
\description{
Get species information table for Phytoplankton and Zooplankton
}
\examples{
dat <- pr_get_SpeciesInfo(Type = "Phytoplankton")
dat <- pr_get_SpeciesInfo(Type = "Zooplankton")
}
