% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_boo_data.R
\name{pr_get_SppCount}
\alias{pr_get_SppCount}
\title{Get the summary plankton observations}
\usage{
pr_get_SppCount(gp)
}
\arguments{
\item{gp}{The group of plankton requested. Either "Zooplankton" or "Phytoplankton"}
}
\value{
a dataframe with a species summary
}
\description{
Get the summary plankton observations from the NRS and CPR.
}
\examples{
df <- pr_get_SppCount("Zooplankton")
}
