% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pr_get_StationName}
\alias{pr_get_StationName}
\title{Add .data$StationName to data}
\usage{
pr_get_StationName(df)
}
\value{
A dataframe with .data$StationName added
}
\description{
Add .data$StationName to data
}
