% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_boo_data.R
\name{pr_get_STI}
\alias{pr_get_STI}
\title{Get STI kernel density for each species}
\usage{
pr_get_STI(df)
}
\arguments{
\item{df}{dataframe as output of pr_get_STI()}
}
\value{
df of STI kernel density for each species
}
\description{
Get STI kernel density for each species
}
\examples{
df <- pr_get_STIdata('Z') \%>\% pr_get_STI()

}
