% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plankton.R
\name{pr_get_STI}
\alias{pr_get_STI}
\title{Get STI kernel density for each species}
\usage{
pr_get_STI(Type = "Zooplankton")
}
\arguments{
\item{Type}{Zooplankton (Copepods) or phytoplankton species}
}
\value{
df of STI kernel density for each species
}
\description{
Get STI kernel density for each species
}
\examples{
df <- pr_get_STI("Phytoplankton")

}
