% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_boo_data.R
\name{pr_get_sti}
\alias{pr_get_sti}
\title{Get data for STI plots of species abundance}
\usage{
pr_get_sti(Type = c("P", "Z"))
}
\value{
df to be sued with pr_plot_sti
}
\description{
Get data for STI plots of species abundance
}
\examples{
df <- pr_get_sti("P")
}
