% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_boo_data.R
\name{pr_get_tsdata}
\alias{pr_get_tsdata}
\title{Access data for timeseries and climatology plots}
\usage{
pr_get_tsdata()
}
\value{
dataframe to use in pr_plot functions
}
\description{
Access data for timeseries and climatology plots
}
\examples{
df <- pr_get_tsdata()
}
