% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planktonr_dat.R
\name{pr_get_type}
\alias{pr_get_type}
\title{Generic function and method for Type}
\usage{
pr_get_type(x)
}
\arguments{
\item{x}{planktonr r data object}
}
\value{
attribute \code{Survey}
}
\description{
Generic function and method for Type
}
