% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planktonr_dat.R
\name{pr_get_variable}
\alias{pr_get_variable}
\title{Generic function and method for Variable}
\usage{
pr_get_variable(x)
}
\arguments{
\item{x}{planktonr r data object}
}
\value{
attribute \code{Variable}
}
\description{
Generic function and method for Variable
}
