% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_indices.R
\name{pr_make_climatology}
\alias{pr_make_climatology}
\title{To produce the climatology for plotting}
\usage{
pr_make_climatology(df, x)
}
\arguments{
\item{df}{output of pr_get_indices}

\item{x}{Year, Month, Day, time period of climatology}
}
\value{
dataframe to use in pr_plot_climate functions
}
\description{
To produce the climatology for plotting
}
\examples{
df <- data.frame(Month = rep(1:12,10), StationCode = 'NSI', Values = runif(120, min=0, max=10))
pr_make_climatology(df, Month)
}
