% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sat.R
\name{pr_match_Altimetry}
\alias{pr_match_Altimetry}
\title{Match data for altimetry}
\usage{
pr_match_Altimetry(df, pr)
}
\arguments{
\item{df}{dataframe containing Latitude, Longitude and Date}

\item{pr}{products from GSLA, GSL, UCUR, UCUR, VCUR, UCUR_MEAN, VCUR_MEAN, single or as a list}
}
\value{
df with product output attached
}
\description{
Match data for altimetry
}
\examples{
df <- head(pr_get_DataLocs("NRS"), 5)
res_spat <- 10
altout <- pr_match_Altimetry(df, pr = "GSLA")
}
