% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis.R
\name{pr_model_data}
\alias{pr_model_data}
\title{Run linear model on station and parameters}
\usage{
pr_model_data(df)
}
\arguments{
\item{df}{A dataframe of NR, CPR or LTM}
}
\value{
A dataframe with the model attribute added
}
\description{
Run linear model on station and parameters
}
\examples{
df <- planktonr::pr_get_EOVs(Survey = "NRS") \%>\%
  dplyr::filter(StationCode != "VBM") \%>\%
  dplyr::filter(Parameters == "PigmentChla_mgm3") \%>\%
  pr_remove_outliers(2) \%>\%
  pr_model_data()
}
