% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_relationships.R
\name{pr_plot_box}
\alias{pr_plot_box}
\title{Simple boxplot function using common NRS colouring}
\usage{
pr_plot_box(dat, y)
}
\arguments{
\item{dat}{Dataframe}

\item{y}{Column name for the y axis}
}
\value{
ggplot object
}
\description{
Note that this function assumes wide data with the data to plot as columns.
}
\examples{
dat <- planktonr::pr_get_NRSMicro('Coastal') \%>\%
tidyr::drop_na(tidyselect::all_of(c("Values", "Parameters"))) \%>\%
dplyr::filter(StationCode \%in\% c("TOP", "BAI")) \%>\%
tidyr::pivot_wider(names_from = "Parameters", values_from = "Values", values_fn = mean)
gg <- pr_plot_box(dat, "Bacterial_Temperature_Index_KD")
}
