% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_boo_plot.R
\name{pr_plot_climate}
\alias{pr_plot_climate}
\title{Plot single climatology}
\usage{
pr_plot_climate(df, Survey = "NRS", Trend = "Month", Scale = "identity")
}
\arguments{
\item{df}{dataframe with specified time period, station code and parameter}

\item{Survey}{CPR or NRS data}

\item{Trend}{specified time period}

\item{Scale}{scale of y axis on plot, whatever scale_y_continuous trans accepts}
}
\value{
a climatology plot
}
\description{
Plot single climatology
}
\examples{
df <- pr_get_indices(Survey = "NRS", Type = "P") \%>\%
        dplyr::filter(parameters == "PhytoBiomassCarbon_pgL")
monthly <- pr_plot_climate(df, "NRS", "Month")

df <- pr_get_indices(Survey = "CPR", Type = "Z") \%>\%
        dplyr::filter(parameters == "ZoopAbundance_m3")
annual <- pr_plot_climate(df, "CPR", "Year")
}
