% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries.R
\name{pr_plot_Climatology}
\alias{pr_plot_Climatology}
\title{Plot single climatology}
\usage{
pr_plot_Climatology(df, Trend = "Month", trans = "identity")
}
\arguments{
\item{df}{dataframe with specified time period, station code and parameter}

\item{Trend}{specified time period}

\item{trans}{scale of y axis on plot, whatever scale_y_continuous trans accepts}
}
\value{
a climatology plot
}
\description{
Plot single climatology
}
\examples{
df <- pr_get_Indices(Survey = "NRS", Type = "Phytoplankton") \%>\%
dplyr::filter(Parameters == "PhytoBiomassCarbon_pgL", StationCode \%in\% c("NSI", "PHB"))

monthly <- pr_plot_Climatology(df, Trend = "Month")

df <- pr_get_Indices(Survey = "CPR", Type = "Zooplankton") \%>\%
        dplyr::filter(Parameters == "ZoopAbundance_m3")
annual <- pr_plot_Climatology(df, Trend = "Year")
}
