% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_boo_plot.R
\name{pr_plot_Climatology}
\alias{pr_plot_Climatology}
\title{Plot single climatology}
\usage{
pr_plot_Climatology(df, Survey = "NRS", Trend = "Month", trans = "identity")
}
\arguments{
\item{df}{dataframe with specified time period, station code and parameter}

\item{Survey}{CPR or NRS data}

\item{Trend}{specified time period}

\item{trans}{scale of y axis on plot, whatever scale_y_continuous trans accepts}
}
\value{
a climatology plot
}
\description{
Plot single climatology
}
\examples{
df <- pr_get_Indices(Survey = "NRS", Type = "P") \%>\%
        dplyr::filter(Parameters == "PhytoBiomassCarbon_pgL", StationCode \%in\% c("NSI", "PHB", "MAI"))

monthly <- pr_plot_Climatology(df, "NRS", "Month")

df <- pr_get_Indices(Survey = "CPR", Type = "Z") \%>\%
        dplyr::filter(Parameters == "ZoopAbundance_m3")
annual <- pr_plot_Climatology(df, "CPR", "Year")
}
