% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_boo_plot.R
\name{pr_plot_CPRmap}
\alias{pr_plot_CPRmap}
\title{Title Sidebar panel plot of selected CPR bioregions}
\usage{
pr_plot_CPRmap(df)
}
\arguments{
\item{df}{dataframe containing CPR bioregions to plot}
}
\value{
is a plotly map of the selected bioregions
}
\description{
Title Sidebar panel plot of selected CPR bioregions
}
\examples{
df <- data.frame(BioRegion = c("Temperate East", "South-west"))
cprmap <- pr_plot_CPRmap(df)
}
