% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries.R
\name{pr_plot_DayNight}
\alias{pr_plot_DayNight}
\title{Visualise diel vertical migration patterns showing day vs night abundances}
\usage{
pr_plot_DayNight(df)
}
\arguments{
\item{df}{dataframe as output of pr_get_DayNight() filtered for one species}
}
\value{
plot of relative day and night abundances
}
\description{
Visualise diel vertical migration patterns showing day vs night abundances
}
\examples{
df <- pr_get_DayNight()
plot <- pr_plot_DayNight(df)
}
