% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_boo_plot.R
\name{pr_plot_EOV}
\alias{pr_plot_EOV}
\title{Policy plot}
\usage{
pr_plot_EOV(
  df,
  EOV = "Biomass_mgm3",
  Survey = "NRS",
  trans = "identity",
  col = "blue",
  labels = "yes"
)
}
\arguments{
\item{df}{dataframe containing timeseries data with Parameters and Values, output of pr_get_PolicyData and pr_get_Coeffs}

\item{EOV}{Essential OCean Variable as a parameter}

\item{Survey}{NRS, CPR or LTM 'Long term monitoring'}

\item{trans}{scale for y axis}

\item{col}{colour selection}

\item{labels}{do you want to print labels on the x axes}
}
\value{
plot of timeseries, anomalies and climatology
}
\description{
Policy plot
}
\examples{
df <- pr_get_PolicyData("CPR") \%>\%
  pr_get_Coeffs()
pr_plot_EOV(df, EOV = "BiomassIndex_mgm3", Survey = "CPR",
      trans = "identity", col = "blue", labels = "no")
}
