% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_boo_plot.R
\name{pr_plot_EOV}
\alias{pr_plot_EOV}
\title{Policy plot}
\usage{
pr_plot_EOV(
  df,
  EOV = "Biomass_mgm3",
  Survey = "NRS",
  trans = "identity",
  pal = "matter",
  labels = "yes"
)
}
\arguments{
\item{df}{dataframe containing timeseries data with parameters and Values, output of pr_get_pol and pr_get_coeffs}

\item{EOV}{Essential OCean Variable as a parameter}

\item{Survey}{NRS, CPR or LTM 'Long term monitoring'}

\item{trans}{scale for y axis}

\item{pal}{colour pallet from cmocean}

\item{labels}{do you want to print labels on the x axes}
}
\value{
plot of timeseries, anomalies and climatology
}
\description{
Policy plot
}
\examples{
df <- data.frame(SampleTime_Local = c("2010-01-01","2010-02-25",
                                      "2010-06-21","2010-04-11","2010-08-05"),
             StationName = "Port Hacking", parameters = "Biomass_mgm3", Values = runif(5, 1, 50),
             fv = runif(5, 1, 50), anomaly = runif(5, 1, 3), Month_Local = runif(5, 1, 6)) \%>\%
             dplyr::mutate(SampleTime_Local = as.Date(SampleTime_Local, format = "\%Y-\%m-\%d"))

plot <-  pr_plot_EOV(df, "Biomass_mgm3", "CPR", "identity", "matter", "yes")
}
