% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries.R
\name{pr_plot_EOVs}
\alias{pr_plot_EOVs}
\title{Essential Ocean Variables plot}
\usage{
pr_plot_EOVs(
  df,
  EOV = "Biomass_mgm3",
  trans = "identity",
  col = "blue",
  labels = TRUE
)
}
\arguments{
\item{df}{dataframe containing timeseries data with Parameters and Values, output of pr_get_EOVs and pr_get_Coeffs}

\item{EOV}{Essential Ocean Variable as a parameter}

\item{trans}{scale for y axis}

\item{col}{colour selection}

\item{labels}{do you want to print labels on the x axes}
}
\value{
plot of timeseries, anomalies and climatology
}
\description{
Essential Ocean Variables plot
}
\examples{
df <- pr_get_EOVs("NRS") \%>\%
  dplyr::filter(StationCode == "PHB") \%>\%
  pr_remove_outliers(2)
pr_plot_EOVs(df, EOV = "Biomass_mgm3",
      trans = "identity", col = "blue", labels = FALSE)
}
