% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_boo_plot.R
\name{pr_plot_Gantt}
\alias{pr_plot_Gantt}
\title{Plot Gantt Chart showing plankton sampling status}
\usage{
pr_plot_Gantt(Survey = "NRS")
}
\arguments{
\item{Survey}{"NRS" or "CPR"}
}
\value{
a ggplot
}
\description{
Plot Gantt Chart showing plankton sampling status
}
\examples{
gg <- pr_plot_Gantt(Survey = "NRS")
gg <- pr_plot_Gantt(Survey = "CPR")
}
