% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries.R
\name{pr_plot_latitude}
\alias{pr_plot_latitude}
\title{Plot for latitudinal data}
\usage{
pr_plot_latitude(df, Fill_NA = FALSE, maxGap = 3)
}
\arguments{
\item{df}{dataframe of latitudinal series}

\item{Fill_NA}{fill in gaps in data}

\item{maxGap}{no of NAs over which to fill gaps}
}
\value{
patchwork object
}
\description{
Plot for latitudinal data
}
\examples{
df <- pr_get_NRSMicro("GO-SHIP")
df <- df \%>\% dplyr::filter(
  Parameters == "Archaea_unique_ASVs",
  SampleDepth_m < 101
)
pr_plot_latitude(df, Fill_NA = TRUE, maxGap = 5)
}
