% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_boo_plot.R
\name{pr_plot_latitude}
\alias{pr_plot_latitude}
\title{PLot for latitudinal data}
\usage{
pr_plot_latitude(df, maxDepth = 100, Fill_NA = FALSE, maxGap = 3)
}
\arguments{
\item{df}{dataframe of latitudinal series}

\item{maxDepth}{depth to show plots to}

\item{Fill_NA}{fill in gaps in data}

\item{maxGap}{no of NAs over which to fill gaps}
}
\value{
patchwork object
}
\description{
PLot for latitudinal data
}
\examples{
df <- pr_get_NRSMicro('GO-SHIP')
df <- df \%>\% filter(.data$Parameters == 'Bacterial_Temperature_Index_KD')
pr_plot_latitude(df, maxDepth = 100, Fill_NA = TRUE, maxGap = 5)
}
