% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_boo_plot.R
\name{pr_plot_NRSEnvContour}
\alias{pr_plot_NRSEnvContour}
\title{Contour plots for depth stratified environmental data}
\usage{
pr_plot_NRSEnvContour(df, Interpolation = TRUE, Fill_NA = FALSE)
}
\arguments{
\item{df}{dataframe from pr_get_NRSEnvContour}

\item{Interpolation}{If TRUE data is interpolated at 5m intervals (DAR = 10m)}

\item{Fill_NA}{to fill in NA's or not via zoo::na_approx, only used when interpolation is TRUE}
}
\value{
a contour plot
}
\description{
Contour plots for depth stratified environmental data
}
\examples{
df <- pr_get_NRSEnvContour("Chemistry") \%>\% dplyr::filter(Parameters == "NOx_umolL",
StationCode \%in\% c('MAI', 'NSI', 'PHB'))
plot <- pr_plot_NRSEnvContour(df, Interpolation = TRUE, Fill_NA = TRUE)
}
