% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_enviro.R
\name{pr_plot_NRSEnvContour}
\alias{pr_plot_NRSEnvContour}
\title{Contour plots for depth stratified environmental data}
\usage{
pr_plot_NRSEnvContour(df, na.fill = TRUE)
}
\arguments{
\item{df}{dataframe from pr_get_NRSEnvContour}

\item{na.fill}{TRUE, FALSE or function like mean to fill in gaps in data}
}
\value{
a contour plot
}
\description{
Contour plots for depth stratified environmental data
}
\examples{
df <- pr_get_NRSEnvContour("Chemistry") \%>\% dplyr::filter(Parameters == "Nitrate_umolL",
StationCode \%in\% c('YON', 'MAI', 'PHB', 'NSI'))
plot <- pr_plot_NRSEnvContour(df, na.fill = TRUE)
}
