% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_enviro.R
\name{pr_plot_NRSEnvContour}
\alias{pr_plot_NRSEnvContour}
\title{Contour plots for depth stratified environmental data}
\usage{
pr_plot_NRSEnvContour(df, Interpolation = TRUE, Fill_NA = FALSE, maxGap = 3)
}
\arguments{
\item{df}{dataframe from pr_get_NRSEnvContour}

\item{Interpolation}{If TRUE data is interpolated at 5m intervals (DAR = 10m)}

\item{Fill_NA}{to fill in NA's or not via zoo::na_approx, only used when interpolation is TRUE}

\item{maxGap}{maximum gap across which to interpolate across NAs}
}
\value{
a contour plot
}
\description{
Contour plots for depth stratified environmental data
}
\examples{
df <- pr_get_NRSEnvContour("Chemistry") \%>\% dplyr::filter(Parameters == "Nitrate_umolL",
StationCode \%in\% c('YON', 'MAI', 'PHB', 'NSI'))
plot <- pr_plot_NRSEnvContour(df, Interpolation = TRUE, Fill_NA = FALSE, maxGap = 3)
}
