% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_maps.R
\name{pr_plot_NRSmap}
\alias{pr_plot_NRSmap}
\title{Sidebar panel plot of selected NRS stations}
\usage{
pr_plot_NRSmap(df, Survey = "NRS")
}
\arguments{
\item{df}{dataframe containing station codes to plot}

\item{Survey}{NRS or Coastal}
}
\value{
a map of the selected stations
}
\description{
Sidebar panel plot of selected NRS stations
}
\examples{
df <- data.frame(StationCode = c("MAI", "PHB"))
pmap <- pr_plot_NRSmap(df)
}
