% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_boo_plot.R
\name{pr_plot_NRSmap}
\alias{pr_plot_NRSmap}
\title{#' Setting the colours for plots using cmocean
#'
#' @param pal is the palette name
#' @param n is the number of colours required
#'
#' @return is a list of colours of length n from palette pal
#' @export
#'
#' @importFrom cmocean cmocean
#'
#' @examples
#' plotCols <- pr_get_PlotCols('matter', 5)
pr_get_PlotCols <- function(pal = "matter", n){
plotCols <- cmocean::cmocean(pal)(n)
return(plotCols)
}
Sidebar panel plot of selected NRS stations}
\usage{
pr_plot_NRSmap(df)
}
\arguments{
\item{df}{dataframe containing station codes to plot}
}
\value{
a map of the selected stations
}
\description{
#' Setting the colours for plots using cmocean
#'
#' @param pal is the palette name
#' @param n is the number of colours required
#'
#' @return is a list of colours of length n from palette pal
#' @export
#'
#' @importFrom cmocean cmocean
#'
#' @examples
#' plotCols <- pr_get_PlotCols('matter', 5)
pr_get_PlotCols <- function(pal = "matter", n){
plotCols <- cmocean::cmocean(pal)(n)
return(plotCols)
}
Sidebar panel plot of selected NRS stations
}
\examples{
df <- data.frame(StationCode = c("NSI", "PHB"))
pmap <- pr_plot_NRSmap(df)
}
