% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_boo_plot.R
\name{pr_plot_NRSmap}
\alias{pr_plot_NRSmap}
\title{Sidebar panel plot of selected NRS stations}
\usage{
pr_plot_NRSmap(df)
}
\arguments{
\item{df}{dataframe containing station codes to plot}
}
\value{
is a plotly map of the selected stations
}
\description{
Sidebar panel plot of selected NRS stations
}
\examples{
df <- data.frame(Code = c("NSI", "PHB"))
pmap <- pr_plot_NRSmap(df)
}
