% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_maps.R
\name{pr_plot_PCImap}
\alias{pr_plot_PCImap}
\title{Map Phytoplankton Colour Index (PCI) from CPR samples around Australia}
\usage{
pr_plot_PCImap(df)
}
\arguments{
\item{df}{dataframe with location and seasonal PCI data}
}
\value{
plot of PCI around Australia
}
\description{
Map Phytoplankton Colour Index (PCI) from CPR samples around Australia
}
\examples{
pr_get_PCIData() \%>\% pr_plot_PCImap()
}
