% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_boo_plot.R
\name{pr_plot_scatter}
\alias{pr_plot_scatter}
\title{Simple function to scatter 2 data columns using common NRS colouring}
\usage{
pr_plot_scatter(df, x, y, Trend = "none")
}
\arguments{
\item{df}{Dataframe}

\item{x}{Column name for the x axis}

\item{y}{Column name for the y axis}

\item{Trend}{Trend line through scatter plot}
}
\value{
ggplot object
}
\description{
Note that this function assumes wide data with the data to plot as columns.
}
\examples{
df <- planktonr::pr_get_NRSMicro() \%>\%
tidyr::drop_na(tidyselect::all_of(c("Values", "Parameters"))) \%>\%
dplyr::filter(.data$StationCode \%in\% c("NSI", "PHB")) \%>\%
tidyr::pivot_wider(names_from = "Parameters", values_from = "Values", values_fn = 'mean')
gg <- pr_plot_scatter(df, "Bacterial_Temperature_Index_KD",
"nitrogen_fixation_organisms", Trend = 'none')
}
