% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_info.R
\name{pr_plot_TaxaAccum}
\alias{pr_plot_TaxaAccum}
\title{Taxa Accumulation Curve}
\usage{
pr_plot_TaxaAccum(dat)
}
\arguments{
\item{dat}{A dataframe of plankton data}
}
\value{
a ggplot object.
}
\description{
Plot a taxa accumulation curve for everything that is identified by the IMOS plankton team
}
\examples{
dat <- pr_get_TaxaAccum(Survey = "NRS", Type = "Zooplankton")
p <- pr_plot_TaxaAccum(dat)
dat <- pr_get_TaxaAccum(Survey = "CPR", Type = "Phytoplankton")
p <- pr_plot_TaxaAccum(dat)
}
