% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_boo_plot.R
\name{pr_plot_timeseries}
\alias{pr_plot_timeseries}
\title{Plot basic timeseries}
\usage{
pr_plot_timeseries(df, Survey = "NRS", pal = "matter", Scale = "identity")
}
\arguments{
\item{df}{dataframe with SampleDate_Local, station code and parameter name and values}

\item{Survey}{CPR or NRS data}

\item{pal}{is the palette name from cmocean}

\item{Scale}{scale of y axis on plot, whatever scale_y_continuous trans accepts}
}
\value{
a timeseries plot
}
\description{
Plot basic timeseries
}
\examples{
df <- data.frame(SampleTime_Local = c("2012-08-21", "2012-09-01", "2012-08-15", "2012-09-18"),
StationCode = 'NSI', parameters = 'Biomass_mgm3', Values = runif(4, min=0, max=10),
Survey = 'NRS')
df <- df \%>\%
      dplyr::mutate(SampleTime_Local = as.POSIXct(paste(SampleTime_Local, "00:00:00"),
      format = "\%Y-\%m-\%d \%H:\%M:\%S"))
timeseries <- pr_plot_timeseries(df, 'NRS', 'matter')
}
