% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_boo_plot.R
\name{pr_plot_timeseries}
\alias{pr_plot_timeseries}
\title{Plot basic timeseries}
\usage{
pr_plot_timeseries(df, Survey = "NRS", pal = "matter", Scale = "identity")
}
\arguments{
\item{df}{dataframe with SampleDate_Local, station code and parameter name and values}

\item{Survey}{CPR or NRS data}

\item{pal}{is the palette name from cmocean}

\item{Scale}{scale of y axis on plot, whatever scale_y_continuous trans accepts}
}
\value{
a timeseries plot
}
\description{
Plot basic timeseries
}
\examples{
df <- pr_get_indices("NRS", "Z") \%>\%
  dplyr::filter(parameters == "Biomass_mgm3")
timeseries <- pr_plot_timeseries(df, "NRS", "matter")
}
