% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_boo_plot.R
\name{pr_plot_TimeSeries}
\alias{pr_plot_TimeSeries}
\title{Plot basic timeseries}
\usage{
pr_plot_TimeSeries(df, Survey = "NRS", trans = "identity")
}
\arguments{
\item{df}{dataframe with SampleDate_Local, station code and parameter name and values}

\item{Survey}{CPR or NRS data}

\item{trans}{scale of y axis on plot, whatever scale_y_continuous trans accepts}
}
\value{
a timeseries plot
}
\description{
Plot basic timeseries
}
\examples{
df <- pr_get_Indices("NRS", "Z") \%>\%
  dplyr::filter(Parameters == "Biomass_mgm3")
timeseries <- pr_plot_TimeSeries(df \%>\% dplyr::filter(Parameters == 'Bacterial_Temperature_Index_KD',grepl('Derwent', StationName)), Survey = "Coastal")
}
