% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_boo_plot.R
\name{pr_plot_timeseries}
\alias{pr_plot_timeseries}
\title{Plot basic timeseries}
\usage{
pr_plot_timeseries(df, pal)
}
\arguments{
\item{df}{dataframe with SampleDateLocal, station code and parameters}

\item{pal}{is the palette name from cmocean}
}
\value{
a timeseries plot
}
\description{
Plot basic timeseries
}
\examples{
df <- data.frame(SampleDateLocal = c("2012-08-21", "2012-09-01", "2012-08-15", "2012-09-18"),
Code = 'NSI', Values = runif(4, min=0, max=10))
df <- df \%>\% mutate(SampleDateLocal = as.POSIXct(paste(SampleDateLocal, "00:00:00"), format = "\%Y-\%m-\%d \%H:\%M:\%S"))
timeseries <- pr_plot_timeseries(df, 'matter')
}
