% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries.R
\name{pr_plot_Trends}
\alias{pr_plot_Trends}
\title{Plot temporal Trends in plankton data}
\usage{
pr_plot_Trends(df, Trend = "Raw", method = "lm", trans = "identity")
}
\arguments{
\item{df}{A dataframe containing the plankton timeseries data.}

\item{Trend}{Over what timescale to fit the Trend - "Raw", "Month" or "Year"}

\item{method}{Any method accepted by \code{geom_smooth()}}

\item{trans}{transformation of y axis on plot, whatever \code{scale_y_continuous()} trans accepts
#'}
}
\value{
a timeseries plot
}
\description{
Plot temporal Trends in plankton data
}
\examples{
df <- pr_get_Indices("SOTS", "Phytoplankton") \%>\%
  dplyr::filter(Parameters == "PhytoAbundance_CellsL") \%>\%
  pr_model_data()
pr_plot_Trends(df, method = "loess", Trend = "Month")
pr_plot_Trends(df, Trend = "Year")
pr_plot_Trends(df, Trend = "Raw")
}
