% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_boo_plot.R
\name{pr_plot_trends}
\alias{pr_plot_trends}
\title{Plot temporal Trends in plankton data}
\usage{
pr_plot_trends(
  df,
  Trend = "Raw",
  Survey = "NRS",
  method = "lm",
  trans = "identity"
)
}
\arguments{
\item{df}{A dataframe containing the plankton timeseries data.}

\item{Trend}{Over what timescale to fit the Trend - "Raw", "Month" or "Year"}

\item{Survey}{"NRS" or "CPR" data}

\item{method}{Any method accepted by \code{geom_smooth()}}

\item{trans}{transformation of y axis on plot, whatever \code{scale_y_continuous()} trans accepts
#'}
}
\value{
a timeseries plot
}
\description{
Plot temporal Trends in plankton data
}
\examples{
df <- pr_get_indices("NRS", "Z") \%>\%
  dplyr::filter(parameters == 'Biomass_mgm3')
pr_plot_trends(df, Trend = "Month", Survey = "NRS")
pr_plot_trends(df, Trend = "Year", Survey = "NRS")
pr_plot_trends(df, Trend = "Raw", Survey = "NRS")
}
