% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_boo_plot.R
\name{pr_plot_trends}
\alias{pr_plot_trends}
\title{Plot temporal trends in plankton data}
\usage{
pr_plot_trends(
  df,
  trend = "Raw",
  survey = "NRS",
  method = "lm",
  pal = "matter",
  y_trans = "identity"
)
}
\arguments{
\item{df}{A dataframe containing the plankton timeseries data.}

\item{trend}{Over what timescale to fit the trend - "Raw", "Month" or "Year"}

\item{survey}{"NRS" or "CPR" data}

\item{method}{Any method accepted by \code{geom_smooth()}}

\item{pal}{is the palette name from \code{cmocean()}}

\item{y_trans}{transformation of y axis on plot, whatever \code{scale_y_continuous()} trans accepts
#'}
}
\value{
a timeseries plot
}
\description{
Plot temporal trends in plankton data
}
\examples{
df <- pr_get_indices("NRS", "Z") \%>\%
  dplyr::filter(parameters == 'Biomass_mgm3')
pr_plot_trends(df, trend = "Month", survey = "NRS")
pr_plot_trends(df, trend = "Year", survey = "NRS")
pr_plot_trends(df, trend = "Raw", survey = "NRS")
}
