% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_boo_plot.R
\name{pr_plot_tsclimate}
\alias{pr_plot_tsclimate}
\title{Combined timeseries and climatology plots}
\usage{
pr_plot_tsclimate(df, Survey = c("CPR", "NRS"), pal, Scale = "identity")
}
\arguments{
\item{df}{data frame with SampleDateLocal, time period and parameter}

\item{Survey}{CPR or NRS data}

\item{pal}{is the palette name from cmocean}

\item{Scale}{scale of y axis on plot, whatever scale_y_continuous trans accepts}
}
\value{
plotly combined plot
}
\description{
Combined timeseries and climatology plots
}
\examples{
df <- data.frame(SampleDateLocal = c("2012-08-21", "2012-09-01", "2012-08-15", "2012-09-18"),
Month = sample(1:12, 4), Year = c(2012, 2013, 2014, 2015),
StationCode = c('NSI', 'NSI', 'PHB', 'PHB'),
Values = runif(4, min=0, max=10))
df <- df \%>\% mutate(SampleDateLocal = as.POSIXct(paste(SampleDateLocal, "00:00:00"),
format = "\%Y-\%m-\%d \%H:\%M:\%S"))
monthly <- pr_plot_tsclimate(df, 'NRS', 'matter')
plotly::ggplotly(monthly)
}
