% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries.R
\name{pr_plot_tsfg}
\alias{pr_plot_tsfg}
\title{Time series plot of functional groups}
\usage{
pr_plot_tsfg(df, Scale = "Actual", Trend = "Raw")
}
\arguments{
\item{df}{dataframe in format of output from pr_get_FuncGroups}

\item{Scale}{y axis scale Actual or Percent}

\item{Trend}{Over what timescale to fit the Trend - "Raw", "Month" or "Year"}
}
\value{
plot of fg timseries
}
\description{
Time series plot of functional groups
}
\examples{
df <- pr_get_FuncGroups("NRS", "Phytoplankton") \%>\%
dplyr::filter(StationCode == 'PHB')
plot <- pr_plot_tsfg(df, "Actual", Trend = 'Month')
plot
}
