% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_maps.R
\name{pr_plot_Voyagemap}
\alias{pr_plot_Voyagemap}
\title{Create map showing CPR voyage tracks and sampling locations}
\usage{
pr_plot_Voyagemap(dat, dats, Country = c("Australia"))
}
\arguments{
\item{dat}{dataframe containing all locations to plot}

\item{dats}{dataframe of sample locations to plot}

\item{Country}{countries to plot on map}
}
\value{
a map of the selected bioregions
}
\description{
Create map showing CPR voyage tracks and sampling locations
}
\examples{
dat <- pr_get_NRSMicro("GO-SHIP")
dats <- dat \%>\% dplyr::slice(1:5000)
voyagemap <- pr_plot_Voyagemap(dat, dats, Country = c("Australia", "New Zealand"))
}
