% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_relabel.R
\name{pr_relabel}
\alias{pr_relabel}
\title{Format a parameter string for plotting}
\usage{
pr_relabel(s, style = "ggplot")
}
\arguments{
\item{s}{A string for reformatting}

\item{style}{The style of plotting the string will be used for}
}
\value{
A reformatted string (plotly) or a reformatted expression call (ggplot)
}
\description{
Format a parameter string for plotting
}
\examples{
pr_relabel("Chla_mgm3", style = "ggplot")

}
