% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_rename.R
\name{pr_rename}
\alias{pr_rename}
\title{Standardise column names in IMOS plankton dataframes}
\usage{
pr_rename(dat)
}
\arguments{
\item{dat}{Dataframe with column names to renamed}
}
\value{
Dataframe with new column names
}
\description{
Standardise column names in IMOS plankton dataframes
}
\keyword{internal}
