% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_rename.R
\name{pr_rename}
\alias{pr_rename}
\title{Standardise column names in IMOS plankton dataframes}
\usage{
pr_rename(df)
}
\arguments{
\item{df}{Dataframe with column names to renamed}
}
\value{
Dataframe with new column names
}
\description{
Standardise column names in IMOS plankton dataframes
}
\examples{
df <- data.frame(LATITUDE = -32, LONGITUDE = 160) \%>\%
            pr_rename()
}
