% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planktonr_dat.R
\name{print.planktonr_dat}
\alias{print.planktonr_dat}
\title{Example of a custom print method.}
\usage{
\method{print}{planktonr_dat}(x, ...)
}
\arguments{
\item{...}{}
}
\description{
Example of a custom print method.
}
