% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{summarise.planktonr_dat}
\alias{summarise.planktonr_dat}
\title{Summarise data in a planktonr_dat object}
\usage{
\method{summarise}{planktonr_dat}(.data, ..., .by = NULL, .groups = "drop")
}
\arguments{
\item{.data}{A \code{planktonr_dat} object.}

\item{...}{<\code{\link[dplyr:dplyr_data_masking]{data-masking}}> Name-value pairs of summary statistics.}

\item{.by}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Optional grouping variables.}

\item{.groups}{Control the grouping structure of the result.}
}
\value{
A \code{planktonr_dat} object with summarised data,
preserving original attributes. The grouping structure depends on \code{.groups}.
}
\description{
This is an S3 method for \code{dplyr::summarise} that ensures
\code{planktonr_dat} attributes are preserved during summarisation.
}
