% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planktonr_dat.R
\name{summarise.planktonr_dat}
\alias{summarise.planktonr_dat}
\title{summarise method for planktonr_dat}
\usage{
summarise.planktonr_dat(data, ..., .by, .groups = "drop")
}
\arguments{
\item{data}{The planktonr data object}

\item{...}{Other possible objects to be passed to the tidyverse}

\item{.by}{Group by}

\item{.groups}{Should the groups be dropped}
}
\description{
summarise method for planktonr_dat
}
