% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_pr.R
\name{theme_pr}
\alias{theme_pr}
\title{Define theme_pr() function}
\usage{
theme_pr(FontSize = 16)
}
\arguments{
\item{FontSize}{use to set font size with base_size}
}
\value{
planktonr theme
}
\description{
Define theme_pr() function
}
\examples{
\dontrun{
ggplot() +
  geom_blank() +
  theme_pr()
}
}
